==> Machine Learning/0.Medical Insurance Data_221109(Wed)-checkpoint.ipynb <==
import pandas as pd

data = pd.read_csv('insurance.csv')

data

data.isnull().sum()

data.describe()

data.duplicated()#중복된 데이터가 있는지 확인할 때 사용하는 함수

data[data.duplicated()]#중복된 데이터가 있는지 확인할 때 사용하는 함수

data[data['charges'] == 1639.5631]

data = data.drop_duplicates()

data.columns

obj_col = ['sex', 'children', 'smoker', 'region']

for col in obj_col :
    print(col, data[col].unique()) # 

for col in obj_col :
    print(col,'\n',data[col].value_counts()) # 칼럼에 대한 각각의 수를 파악할 수 있음
    print('-----------')

data.groupby('smoker')['charges'].median() #흡연자별 보험료

data.groupby('region')['charges'].median() #지역별 보험료

q1 = data['charges'].quantile(0.25) #제 1사분위수

q3 = data['charges'].quantile(0.75) #제 3사분위수

outlier = q3 + (q3 - q1) * 1.5 #이상치 값 구할 때 

data[data['charges'] > outlier]

data[data['charges'] <= outlier]

data['charges'].mean()

data['charges'].median()

from sklearn.preprocessing import LabelEncoder

#문자열->숫자로 바꿔 줌
col_list = ['sex', 'smoker', 'region']
for col in col_list :
    label = LabelEncoder()
    label.fit(data[col].unique())
    data[col] = label.transform(data[col])

data

from sklearn.model_selection import train_test_split
from sklearn.linear_model import LinearRegression
from sklearn.metrics import mean_absolute_error

x_train, x_test, y_train, y_test = train_test_split(x, y, test_size = 0.2, randeom_state = 0)

model = LinearRegression()

#train

model.fit(x_train, y_train)

pre_train = model.predict(x_train)

mean_absolute_error(pre_train, y_train)

model.score(x_train, y_train)

#test

pre_test = model.predict(x_test)

mean_absolute_error(pre_test, y_test)

model.score(x_test, y_test)

from sklearn.metrics import accuracy_score

pre_test = model.predict(x_test)

mean_absolute_error(pre_test, y_test)

accuracy_score(model.predict(x_test), y_test)

model.score(x_test, y_test)

#LabelEncoder까지 해준 후 train_test_split 하기 전에 이 부분을 해주어야 한다!!!!!!!!!
from sklearn.preprocessing import PolynomialFeatures

pol = PolynomialFeatures(degree = 2)

x_pol = pol.fit_transform(x)

pol.get_feature_names()

x_pol[0]

----

# Medical Insurance Data
## Medical Insurance 데이터를 이용해서 보험료('charges') 예측하기

모델을 바꾸거나(=좋은 모델을 사용해야 하거나), 다항회귀로 바꾸어주어야한다.

data.info()

data.describe()

data.isnull().sum()

import seaborn as sns
import matplotlib.pyplot as plt

corr = data.corr().round(2)

plt.figure(figsize = (10, 10))
sns.heatmap(corr, annot = True)

import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
from matplotlib import style
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LinearRegression

df = pd.read_csv('insurance.csv')

df.head()

df.shape

df.info()

df.isnull().sum()

df.columns

df.describe()

plt.figure(figsize = (5, 5))
style.use('ggplot')
sns.countplot(x = 'sex', data = df)
plt.title('Gender Distribution')
plt.show()

plt.figure(figsize = (5, 5))
sns.countplot(x = 'smoker', data = df)
plt.title('Smoker')
plt.show()

plt.figure(figsize = (5, 5))
sns.countplot(x = 'region', data = df)
plt.title('Region')
plt.show()

plt.figure(figsize = (5, 5))
sns.barplot(x = 'region', y = 'charges', data = df)
plt.title('Cost vs Region')

plt.figure(figsize=(5,5))
sns.barplot(x='smoker', y='charges', data = df)
plt.title('Charges for smokers')

plt.figure(figsize=(5,5))
sns.barplot(x='sex', y='charges', hue = 'smoker', data = df)
plt.title('Charges for smokers')

fig, axes = plt.subplots(1,3, figsize=(15,5), sharey=True)
fig.suptitle('Visualizing categorical columns')
sns.boxenplot(x='smoker', y='charges', data=df, ax=axes[0])
sns.boxenplot(x='sex', y='charges', data=df, ax=axes[1])
sns.boxenplot(x='region', y='charges', data=df, ax=axes[2])

df[['age', 'bmi', 'children', 'charges']].hist(bins = 30, figsize = (10, 10), color = 'blue')

df.head()

df['sex'] = df['sex'].apply({'male':0, 'female':1}.get)
df['smoker'] = df['smoker'].apply({'yes':1, 'no':0}.get)
df['region'] = df['region'].apply({'southwest':1, 'southeast':2, 'northwest':3, 'northeast':4}.get)

df.head()

plt.figure(figsize = (10, 7))
sns.heatmap(df.corr(), annot = True)
plt.show()

X = df.drop(['charges', 'sex'], axis = 1)
y = df.charges
