==> Machine Learning/1.Comprehension & partial_221115(Tue).ipynb <==
# 오전 수업

# Python Comprehension

- List(리스트)
- Set(집합)
- Dict(딕셔너리)

### List Comprehension: 반복되거나 특정 조건을 만족하는 리스트를 보다 쉽게 만들어 내기 위한 방법

[출력표현식 for 요소 in 입력Sequnce [if 조건식]]

[2 * i for i in range(1, 10)]

# Quiz1
[Quiz1] List Comprehension을 이용해서 1 ~ 10까지 제곱 출력하기
[1, 4, 9, 16, 25, 36, 49, 64, 81, 100]

1 * 1
2 * 2
3 * 3
4 * 4
5 * 5
6 * 6
7 * 7
8 * 8
9 * 9
10 * 10

[i * i for i in range(1, 11)] #Quiz1 정답

# Quiz2
[Quiz2] 1 ~ 20까지 짝수만 다 넣기

ret = []
for i in range(1, 21) :
    if i % 2 == 0 :
        ret.append(i)

ret #Quiz2 정답

# Quiz3
[Quiz3] List Comprehension을 이용해서 oldlist에서 숫자만 뽑아서 제곱해서 newlist를 만들어라.
예상결과: newlist = [1, 4, 9]

oldlist = [1, 2, 'A', False, 3]

[1, 2, 'A', False, 3] -> [1, 2, 3] -> [1 * 1, 2 * 2, 3 * 3] -> [1, 4, 9]

[i * i for i in (oldlist) if type(i) == int]

newlist = [i * i for i in oldlist if type(i) == int] #Quiz3 정답

---

# 혼자 연습

# 2, 4, 6, 8, 10, 12, 14, 16, 18, 20
a = []
for i in range(1, 11) :
    print(2 * i)
    a.append(i)
    print(a)

for문 활용하여 구구단 2단 만들기

2 * 1
2 * 2
2 * 3
2 * 4
2 * 5
2 * 6
2 * 7
2 * 8
2 * 9

for i in range(1, 10) :
    a = (2 * i)
b = [a]
print(b)

for i in range(2, 10) :
    for j in range(1, 10):
        print(i * j)

----

# 혼자 List comprehension 연습

[i for i in range(10)]

a = []
for i in range(10) :
    a.append(i)

a

[i for i in range(10)]

a = []
for i in range(10) :
    a.append(i)
    print(a)

[i * 3 for i in range(1, 10)]

a = []
for i in range(1, 10) :
    a.append(3 * i)

a

[i for i in range(1, 30) if i % 2 == 0] #0~29 중 짝수인 값만 출력하시오. = 2의 배수

a = []
for i in range(1, 30) :
    if i % 2 == 0 :
        print(i)
        a.append(i)

a

#1 ~ 27에서 3으로 나누어 주었을 때, 나머지가 0인 값인 애들을 출력하시오. = 3의 배수
[i for i in range(1, 28) if i % 3 == 0]

# 예제1) 2개의 반복문 사용하기
a = ['a', 'b', 'c', 'd', 'e']
b = ['1', '2', '3', '4', '5']

[i + j for i in a for j in b]

# 2개의 반복문 사용하기
c = []
for i in a :
    for j in b:
        c.append(i+j)

c

# 2개의 조건문 사용하기
for i in range(50) :
    if i % 2 == 0 :
        print(i, end = ' ')
        print(type(i))

# 2개의 조건문 사용하기
for i in range(50) :
    if i % 2 == 0 :
        if i % 3 == 0 :
            print(i, end = ' ')
            print(type(i))

[i for i in range(50) if i % 2 == 0 if i % 3 == 0] # 2개의 조건문 사용하기

# 조건문에서 else 사용하기 *****
['even' if i % 2 == 0 else 'odd' for i in range(10)]

for i in range(10) :
    print(i)
    if i % 2 == 0 :
        print('even', end = ' ')
    else :
        print('odd', end = ' ')

# 조건문에서 elif 사용하기 ***** 왜 i가 0일 때 'zero'가 출력이 안되나? 앞에 조건(if i % 2 == 0 조건에 만족하여, 여기에 해당하는 'even'을 출력하게 되므로 'zero'가 출력이 되지 않는다.
['even' if i % 2 == 0 / else 'zero' if i == 0 / else 'odd' for i in range(10)]

# 조건문에서 elif 사용하기 *****
for i in range(10) :
    if i % 2 == 0 :
        print('even', end = ' ')
    elif i == 0 :
        print('zero', end = ' ')
    else :
        print('odd', end = ' ')

['zero' if i == 0 else 'even' if i % 2 == 0 else 'odd' for i in range(10)]

for i in range(10) :
    if i == 0 :
        print('zero', end = ' ')
    elif i % 2 == 0 :
        print('even', end = ' ')
    else :
        print('odd', end = ' ')

['one' if i == 1 / else 'two' if i == 2 / else 'three' if i == 3 / else 'four' if i == 4 / else 'hum' 
 for i in range(10)]

for i in range(10) :
    if i == 1 :
        print('one', end = ' ')
    elif i == 2 :
        print('two', end = ' ')
    elif i == 3 :
        print('three', end = ' ')
    elif i == 4:
        print('four', end = ' ')
    else :
        print('hum', end = ' ')

---

# Set Comprehension (집합)

oldlist = [1, 1, 2, 3, 3, 4]

{i * i for i in oldlist}

# 혼자 연습

oldlist = [1, 1, 1, 1, 1,1,1,1,1,1,1,1,1, 2, 3, 4, 4, 4, 4, 4, 4, 4, 4, 5, 5, 5, 5, 5]

{i for i in oldlist}

---

# Dictionary Comprehension (딕셔너리) *****

{Key:Value for 요소 in 입력Sequence [if 조건식]}

id_name = {1:'남성헌', 2:'고상현', 3:'윤승현'}

이름 -> key / 숫자 -> value

a = {value : key for key, value in id_name.items()}

a

# 혼자 다시 연습

name_id = {1:'남성현', 2:'고상현', 3:'윤승현'}

name_id

name_id.items()

#이름을 key, 숫자를 value로 바꾸시오.
a = {value : key for key, value in name_id.items()}

a

#Dict comprehension 예제 1) 
students = ['철수', '영희', '길동', '순희']
{student: 0 for student in students}

#Dict comprehension 예제 1-1) 
students = ['철수', '영희', '길동', '순희']
{student: _ for student, _ in (students)}

#Dict comprehension 예제 1-1) 
fruits = ['사과', '포도', '자두', '수박']
{fruit: _ for fruit, _ in (fruits)}

#Dict comprehension 예제 1-2) 각 학생들 순서 매기기
students = ['철수', '영희', '길동', '순희']
{student: index for index, student in enumerate(students, start= 1)}

#Dict comprehension 예제 1-2) 각 학생들 순서 매기기
a = {}
for index, student in enumerate(students, start = 1) :
    a[student] = index

a

# 전학 간 학생 제거
scores = {'철수': 50, '영희': 80, '길동':90, '순희': 60, '전학생':100}
scores = {name: score for name, score in scores.items() if name != '전학생'}
print(scores)

# 전학 간 학생 제거
for name, score in scores.items() :
    if name != '전학생' :
        print({name : score}, end = ',')
        scores[name] = score

scores

# 통과, 불통과
grades = {name: 'PASS' if value > 70 else 'No-PASS' for name, value in scores.items()}

grades

for name, value in scores.items() :
    if value > 70 :
        print({name: 'PASS'}, end = ',')
    else :
        print({name: 'No-PASS'}, end = ',')
        scores[name] = value

scores

---

# 혼자 연습

id_name.items()

for i, j in id_name.items() :
    print(i, j)
    new_id_name = {j: int(i)}
    print(new_id_name)
a = new_id_name.items()

# 다시
name_id = {}
for key, value in id_name.items() :
    print(key, value)
    name_id[value] = key
    print(name_id)

name_id

---

name_id = {1:'남성헌', 2:'고상현', 3:'윤승현'}

name_id

a = {key : val for val, key in name_id.items()}

a

a.keys()

a.values()

[key for key, val in a.items() if val == 2]

[key for key, val in a.items() if val == 2][0]

for key, val in a.items() :
    if val == 2 :
        print(key)

# List Comprehension을 사용해서 name_id에서 val가 2인 key를 출력하시오.
[key for key, val in a.items() if val == 2]
[key for key, val in a.items() if val == 2][0]

---

{Key:Value for 요소 in 입력Sequence [if 조건식]}

# Quiz
[Quiz] Comprehension을 이용하여 1부터 10까지 홀수를 key로 하고, value를 key의 제곱으로 하는 Dict 만들기

a = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]

b = {1:1*1, 2:2*2, 3:3*3, 4:4*4, 5:5*5, 6:6*6, 7:7*7, 8:8*8, 9:9*9, 10:10*10}

for i in a :
    if i % 2 != 0 :
        print(i)
        print(i * i)

{i : i * i for i in a  if i % 2 != 0}

---

# Partial
## 매개 변수가 1개인 함수에 여러 매개변수를 전달해야 할 때 많이 사용

def double(x) :
    return x ** 2

def triple(x) :
    return x ** 3

print(double(2))
print(triple(2))

def multiple(x, y) :
    return x ** y

multiple(2, 2)

multiple(2, 3)

from functools import partial

dbl = partial(multiple, y = 2)
tpl = partial(multiple, y = 3)

dbl(2)

tpl(3)

def multiple(x, y, z) :
    return x ** y ** z

dbl = partial(multiple, y = 2, z = 2)
tpl = partial(multiple, y = 3, z = 3)

dbl(2)

tpl(3)
