==> Machine Learning/1.eleveinfo_221103(목)-checkpoint.ipynb <==
11/07(월) 시험 -> 데이터 모델링 수업 진행 예정

## 주소1, 주소2, 건물명, 승강기번호, 
# 승강기상태, 최종검사일, 적재하중, 제조업체전화번호

# 데이터 전체 200개 불러오기

import requests
from bs4 import BeautifulSoup
import pandas as pd

url = ('http://apis.data.go.kr/openapi/service/ElevatorInformationService/getElevatorList?serviceKey=BAc5c%2B7CgdSM2dp9F0gq%2FN7FR4TwTghq%2BzHtk7BmmX%2B588SuKYRFAW77xC6eaS6kpEZgWxZnz9a1TSk%2FO%2F6gUg%3D%3D&numOfRows=10&pageNo=1&sido=%EC%84%9C%EC%9A%B8%ED%8A%B9%EB%B3%84%EC%8B%9C&sigungu=%EC%9D%80%ED%8F%89%EA%B5%AC')

req = requests.get(url)

soup = BeautifulSoup(req.text, 'xml')

soup

address1 = soup.select('address1')
address2 = soup.select('address2')
buldNm = soup.select('buldNm')
elevatorNo = soup.select('elevatorNo')
elvtrStts = soup.select('elvtrStts')
lastInspctDe = soup.select('lastInspctDe')
liveLoad = soup.select('liveLoad')
a = soup.select('manufacturerTelno')

add1

data = []
for i in range(len(address1)) :
    temp = []
    add1 = soup.select('address1')
    add2 = soup.select('address2')
    building_name = soup.select('buldNm')
    ele_num = soup.select('elevatorNo')
    ele_state = soup.select('elvtrStts')
    final_inspect = soup.select('lastInspctDe')
    live_load = soup.select('liveLoad')
    company_num = soup.select('manufacturerTelno')

data = []
for i in range(9) :
    temp = []
    temp.append(address1[i].text)
    temp.append(address2[i].text)
    temp.append(buldNm[i].text)
    temp.append(elevatorNo[i].text)
    temp.append(elvtrStts[i].text)
    temp.append(lastInspctDe[i].text)
    temp.append(liveLoad[i].text)
    temp.append(a[i].text)
    data.append(temp)



data

from tqdm import tqdm #for문을 얼마나 진행했는지 확인 가능

data = []
col_names = []
for m in tqdm(range(6, 11)) :
    url = 'http://apis.data.go.kr/openapi/service/ElevatorInformationService/getElevatorList'
    serviceKey='BAc5c%2B7CgdSM2dp9F0gq%2FN7FR4TwTghq%2BzHtk7BmmX%2B588SuKYRFAW77xC6eaS6kpEZgWxZnz9a1TSk%2FO%2F6gUg%3D%3D'
    numOfRows='10'
    pageNo='1'
    sido='%EC%84%9C%EC%9A%B8%ED%8A%B9%EB%B3%84%EC%8B%9C'
    sigungu='%EC%9D%80%ED%8F%89%EA%B5%AC'    
    
    api_url = url + '?' + 'serviceKey=' + serviceKey + '&numOfRows=' + numOfRows \
    + '&pageNo=' + pageNo + '&sido=' + sido + '&sigungu=' + sigungu

    req = requests.get(api_url)
    soup = BeautifulSoup(req.text, 'xml')

df = pd.DataFrame(data, columns = col_names)

df

url = 'http://apis.data.go.kr/openapi/service/ElevatorInformationService/getElevatorList'
    
serviceKey='BAc5c%2B7CgdSM2dp9F0gq%2FN7FR4TwTghq%2BzHtk7BmmX%2B588SuKYRFAW77xC6eaS6kpEZgWxZnz9a1TSk%2FO%2F6gUg%3D%3D'
numOfRows='10'
pageNo='1'
sido='%EC%84%9C%EC%9A%B8%ED%8A%B9%EB%B3%84%EC%8B%9C'
sigungu='%EC%9D%80%ED%8F%89%EA%B5%AC'    
    
api_url = url + '?' + 'serviceKey=' + serviceKey + '&numOfRows=' + numOfRows \
+ '&pageNo=' + pageNo + '&sido=' + sido + '&sigungu=' + sigungu

api_url

data = []
col_names = []
for m in tqdm(range(6, 11)) :
    url = 'http://apis.data.go.kr/openapi/service/ElevatorInformationService/getElevatorList
    serviceKey='BAc5c%2B7CgdSM2dp9F0gq%2FN7FR4TwTghq%2BzHtk7BmmX%2B588SuKYRFAW77xC6eaS6kpEZgWxZnz9a1TSk%2FO%2F6gUg%3D%3D'
    numOfRows='10'
    pageNo='1'
    sido='%EC%84%9C%EC%9A%B8%ED%8A%B9%EB%B3%84%EC%8B%9C'
    sigungu='%EC%9D%80%ED%8F%89%EA%B5%AC'    
    
    api_url = url + '?' + 'serviceKey=' + serviceKey + '&numOfRows=' + numOfRows \
    + '&pageNo=' + pageNo + '&sido=' + sido + '&sigungu=' + sigungu

    req = requests.get(api_url)
    soup = BeautifulSoup(req.text, 'xml')
    
    if not col_names :
        for item in soup.items()[0] :
            col_names.append(item.name)
    
    for item in soup.items :
        temp = []
        for t in item :
            temp.append(t.text.strip())
        data.append(temp)

#강사님 코드

import requests
from bs4 import BeautifulSoup
import pandas as pd

import warnings
warnings.filterwarnings('ignore')

url = 'http://apis.data.go.kr/openapi/service/ElevatorInformationService/getElevatorList'
serviceKey = 'BAc5c+7CgdSM2dp9F0gq/N7FR4TwTghq+zHtk7BmmX+588SuKYRFAW77xC6eaS6kpEZgWxZnz9a1TSk/O/6gUg=='
numOfRows = '20'
pageNo = '1'
sido= '서울특별시' #'%EC%84%9C%EC%9A%B8%ED%8A%B9%EB%B3%84%EC%8B%9C'
sigungu= '은평구' #'%EC%9D%80%ED%8F%89%EA%B5%AC'  

params = {'serviceKey': serviceKey, 'numOfRows': numOfRows, 'pageNo': pageNo, 'sido': sido, 'sigungu': sigungu}

req = requests.get(url, params = params, verify = False)

soup = BeautifulSoup(req.text, 'lxml')

col_list = ['address1', 'address2','buldNm', 'elevatorNo', 'elvtrStts', 'lastInspctDe', 'liveLoad', 'manufacturerTelno']

data = []
for col in col_list :
    temp = []
    for d in soup.select(col) :
        temp.append(d.text)
    data.append(temp)

import numpy as np

df = pd.DataFrame(data).T

df.columns = col_list

df

for col in col_list :
    print(col, len(soup.select(col)))

# 위 데이터 프레임은 19번째 인덱스에서의 'manufacturerTelno'가 None이 되어 데이터 값이 날라감. 그래서 망한 데이터 프레임임.

# 망한 데이터 프레임을 다시 살리는 방법

for item in soup.items :
    print(item)
    print('------')

soup.item.select('address1')

#방법2

col_list = ['address1', 'address2','buldNm', 'elevatorNo', 'elvtrStts', 'lastInspctDe', 'liveLoad', 'manufacturerTelno']

col_list = ['address1', 'address2','buldNm', 'elevatorNo', 'elvtrStts', 'lastInspctDe', 'liveLoad', 'manufacturerTelno']
data = []
for item in soup.items :
    temp = []
    for col in col_list :
        try :
            temp.append(item.select(col)[0].text)
        except :
            temp.append('')
    data.append(temp)

df = pd.DataFrame(data, columns = col_list)

df

#'예외처리' 하는 방법: try / except

a = 1
b = '2'
try :
    print('try 시작')
    c = a + b
    print('try 끝')
except :
    print('except 시작')
    c = a + int(b)
print(c)

a = 1
b = 2
try :
    print('try 시작')
    c = a + b
    print('try 끝')
except :
    print('except 시작')
    c = a + int(b)
print(c)

b = input()

a = 1
try :
    c = a + int()
    

#방법1

data = []
for item in soup.items :
    temp = []
    for v in item :
        temp.append(v.text.strip())
    data.append(temp)

col_list = []
for i in soup.item :
    col_list.append(i.name)

df = pd.DataFrame(data, columns = col_list)

df.iloc[9]

url = 'http://apis.data.go.kr/openapi/service/ElevatorInformationService/getElevatorList'
serviceKey = '0kYXnKBN3P1BPz+PbMj53+Xf9Lasd9mZvYohzWDq+10Tagyqazw0pf62r0VMYgKLZZednlK4k5uPA6V49OvEcw=='
numOfRows = '10'
pageNo = '2'
sido = '서울특별시'
sigungu = '은평구'

params = {'serviceKey' : serviceKey, 'numOfRows' : numOfRows, 'pageNo' : pageNo, 'sido':sido, 'sigungu':sigungu }
req = requests.get(url, params=params,verify=False)
soup = BeautifulSoup(req.text,'lxml')
col_list = ['address1', 'address2', 'buldNm', 'elevatorNo', 'elvtrStts','lastInspctDe', 'liveLoad', 'manufacturerName', 'manufacturerTelno']
data = []

for item in soup.items :
    temp = []
    for col in col_list :
        try :
            temp.append(item.select(col)[0].text)
        except :
            temp.append('')
    data.append(temp)

## 주소1, 주소2, 건물명, 승강기번호, 
# 승강기상태, 최종검사일, 적재하중, 제조업체전화번호

# 데이터 전체 200개 불러오기

data = []
for page in range(1, 3): 
    url = 'http://apis.data.go.kr/openapi/service/ElevatorInformationService/getElevatorList'
    serviceKey = '0kYXnKBN3P1BPz+PbMj53+Xf9Lasd9mZvYohzWDq+10Tagyqazw0pf62r0VMYgKLZZednlK4k5uPA6V49OvEcw=='
    numOfRows = '100'
    pageNo = page
    sido = '서울특별시'
    sigungu = '은평구'

    params = {'serviceKey' : serviceKey, 'numOfRows' : numOfRows, 'pageNo' : pageNo, 'sido':sido, 'sigungu':sigungu }
    req = requests.get(url, params=params,verify=False)
    soup = BeautifulSoup(req.text,'lxml')
    col_list = ['address1', 'address2', 'buldNm', 'elevatorNo', 'elvtrStts','lastInspctDe', 'liveLoad', 'manufacturerTelno']
    
    for item in soup.items :
        temp = []
        for col in col_list :
            try :
                temp.append(item.select(col)[0].text)
            except :
                temp.append('')
        data.append(temp)

df = pd.DataFrame(data, columns = col_list)

df.shape

df.to_csv('elevInfo.csv', index = False, encoding='euc-kr')

df
