==> Machine Learning/1.sonar_221109(Wed)-checkpoint.ipynb <==
# 수중 음파 탐지기(sonar) 데이터 셋
## 문제: Sonar 데이터 셋을 이용해서 Mines와 Rocks를 구분하는 모델 만들기

import pandas as pd
from sklearn.linear_model import LogisticRegression #Classification의 한 종류
from sklearn.model_selection import train_test_split

sonar = pd.read_csv('sonar_csv.csv')

sonar

sonar.info()

sonar.isnull().sum()

sonar.describe()

sonar.duplicated() #중복된 데이터가 있는지 확인할 때 사용하는 함수

sonar[sonar.duplicated()] #여기에서는 중복된 데이터가 없다는 것을 확인할 수 있음

sonar['Class'].unique()

sonar.columns

obj_col = ['Class']

for col in obj_col :
    print(col, sonar[col].unique())

for col in obj_col :
    print(col, sonar[col].value_counts())

from sklearn.preprocessing import LabelEncoder #문자열->숫자로 바꾸어줄 때 쓰는 패키치 내부 함수

#문자열 -> 숫자 / Rock: 1, Mine: 0
col_list = ['Class']
for col in col_list :
    label = LabelEncoder()
    label.fit(sonar[col].unique())
    sonar[col] = label.transform(sonar[col])

sonar

from sklearn.model_selection import train_test_split
from sklearn.linear_model import LinearRegression #회귀할 때 사용
from sklearn.metrics import mean_absolute_error

x = sonar.drop('Class', axis = 1)

y = sonar['Class']

x_train, x_test, y_train, y_test = train_test_split(x, y, test_size = 0.1, stratify = y, random_state = 0)
# stratify =  -> 기준점

print(x_train.shape, y_train.shape) #학습시킬 문제, 학습시킬 문제에 대한 정답
print(x_test.shape, y_test.shape) #시험문제, 시험문제에 대한 정답

# LogisticRegression(기존) 모델로 돌렸을 경우

model = LogisticRegression() #모델 선언

# train

model.fit(x_train, y_train) # 공부(학습)시킬 때 필요한 함수

pre_train = model.predict(x_train) #트레인에 대한 예측값, pre_train

mean_absolute_error(pre_train, y_train)

model.score(x_train, y_train)

# test

pre_test = model.predict(x_test) #트레인에 대한 예측값, pre_train

mean_absolute_error(pre_test, y_test)

model.score(x_test, y_test)

# LogisticRegressionCV 모델로 돌렸을 경우 (참고 -> CV: Cross Validation)

from sklearn.linear_model import LogisticRegressionCV #분류에 사용한다.

model2 = LogisticRegressionCV() # model2 = LogisticRegressionCV(max_iter=100, cv=1)

# train

model2.fit(x_train, y_train)

model2.score(x_train, y_train)

# test

model2.score(x_test, y_test)

sensor = x_train.iloc[0].values.reshape(1, -1)

model2.predict(sensor)

if model2.predict(sensor) == 1 :
    print('mine')
    print(model2.predict(sensor) + '이 탐지되었습니다.')

y_train.iloc[0]

pip list

---

# 스케일러(Scaler) / 스케일러 할때는 이상치 값을 제거해주는게 정석이다.

Standard Scaler: 기본 스케일, 평균과 표준편차 사용
MinMax Scaler: 최대/최솟값이 각각 1, 0이 되도록 스케일링
MaxAbsScaler: 최대절대값과 0이 각각 1, 0이 되도록 스케일링
RobustScaler: 

import pandas as pd
from sklearn.linear_model import LogisticRegression #Classification의 한 종류
from sklearn.model_selection import train_test_split



sonar = pd.read_csv('sonar_csv.csv')

sonar

for c in sonar.drop('Class', axis=1).columns :
    print(c)
    print(sonar[c].mean())
    print(sonar[c].median())
    print('---------------')

x = sonar.drop('Class', axis = 1)

y = sonar['Class']

x_train, x_test, y_train, y_test = train_test_split(x, y, test_size = 0.1, stratify = y, random_state = 0) # stratify =  -> 기준점

from sklearn.preprocessing import StandardScaler

scaler = StandardScaler() # 평균과 표준편차 

scaler.fit(x_train) # x_train의 평균과 표준편차 / = scaler.fit_transform(x_train)
x_test = scaler.transform(x_test)

scaler.fit(x_train) # 속성별로 학습시키고
x_train = x_train.transform(x_train) # 속성별 내부에서의 각각의 값들을 StandardScaler() 해주는 과정

-> # x_train = scaler.fit_transform(x_train)

x_train = scaler.fit_transform(x_train)
x_test= scaler.transform(x_test)
# x_test = scaler.transform(x_test)

x_train

model3 = LogisticRegression()

model3.fit(x_train, y_train)

model3.predict(x_train)

print(x_train.shape, y_train.shape)

model3.score(x_train, y_train)

model3.score(x_test, y_test)

real = x.iloc[0].values

real

# 1. 데이터 -> 전처리(scale) -> train/test (에이스 version) -> 학습(fit) -> 검증(score)
# 1-1. 데이터 -> train/test -> scale (강사님 version) -> 학습(fit) -> 검증(score)

# real -> 전처리(scale) -> 예측(predict)

real = x.iloc[0].values.reshape(1, -1)

real = scaler.transform(real)

model3.predict(real)

predict = model3.predict(real)[0]

print(predict + '이 탐지되었습니다.')

import pickle

with open('scaler.pkl', 'wb') as f :
    pickle.dump(scaler, f)

with open('model3.pkl', 'wb') as f :
    pickle.dump(model3, f)
