==> Machine Learning/1-1.loan model(self review)_221111(Fri)-checkpoint.ipynb <==
import pandas as pd
import numpy as np
import warnings
warnings.filterwarnings('ignore')
import pickle

from sklearn.svm import SVC
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import LabelEncoder

loan_data = pd.read_csv('loan.csv')

loan_data

loan_data.dtypes

#전처리

#object의 칼럼들만 리스트로 따로 모아라.
obj_col = []
for i, j in zip(loan_data.dtypes.index, loan_data.dtypes.values) :
    if j == 'object' :
        obj_col.append(i)

obj_col

for col in obj_col[1:] :
    print(loan_data[col].unique())

null_col = ['Gender', 'Married', 'Dependents', 'Self_Employed']

# 찍어보기 위함
for col in null_col :
    print(loan_data[col].value_counts(dropna=False))

for col in null_col :
    loan_data[col].fillna(loan_data[col].mode()[0], inplace=True)

loan_data['LoanAmount'].fillna(loan_data['LoanAmount'].mode()[0], inplace = True)

loan_data['Loan_Amount_Term'].fillna(loan_data['Loan_Amount_Term'].mode()[0], inplace = True)

loan_data['Credit_History'].fillna(loan_data['Credit_History'].mode()[0], inplace = True)

from sklearn.preprocessing import LabelEncoder

label = LabelEncoder()

#
for col in obj_col[1:] :
    label = LabelEncoder()
    loan_data[col] = label.fit_transform(loan_data[col])

---

null_col = ['Gender', 'Married', 'Dependents', 'Self_Employed']
for col in null_col :
    loan_data[col].fillna(loan_data[col].mode()[0], inplace=True)
    
loan_data['LoanAmount'].fillna(loan_data['LoanAmount'].mode()[0], inplace = True)
loan_data['Loan_Amount_Term'].fillna(loan_data['Loan_Amount_Term'].mode()[0], inplace = True)
loan_data['Credit_History'].fillna(loan_data['Credit_History'].mode()[0], inplace = True)

obj_col = ['Gender', 'Married', 'Dependents','Education','Self_Employed','Property_Area']

---

# 모델 학습

X = loan_data.drop(['Loan_ID', 'Loan_Status'], axis = 1)
Y = loan_data['Loan_Status']

x_train, x_test, y_train, y_test = train_test_split(X, Y, test_size=0.2, stratify=Y, random_state=0)

model = SVC()

model

model.fit(x_train, y_train)

model.score(x_train, y_train)

model.score(x_test, y_test)

from sklearn.preprocessing import MinMaxScaler

scaler = MinMaxScaler()

x_train = scaler.fit_transform(x_train)

with open('scaler.pkl', 'wb') as f :
    pickle.dump(scaler, f)

x_test = scaler.transform(x_test)

model2 = SVC()

model2.fit(x_train, y_train)

model2.score(x_train, y_train)

model2.score(x_test, y_test)

#모델 저장

with open('model2.pkl', 'wb') as f :
    pickle.dump(model2, f)

----

Real system에 필요한 전처리 코드

import pandas as pd
import numpy as np
import warnings
warnings.filterwarnings('ignore')
import pickle

from sklearn.svm import SVC
from sklearn.model_selection import train_test_split

loan_data = pd.read_csv('loan.csv')

#1

null_dict = {}

null_col = ['Gender', 'Married', 'Dependents', 'Self_Employed']
for col in null_col :
    null_dict[col] = loan_data[col].mode()[0]
    
null_dict['LoanAmount'] = loan_data['LoanAmount'].median()
null_dict['Loan_Amount_Term'] = loan_data['Loan_Amount_Term'].mode()[0]
null_dict['Credit_History'] = loan_data['Credit_History'].mode()[0]

#바로 위에 코드랑 같은 코드임
null_dict = {}

null_col = ['Gender', 'Married', 'Dependents', 'Self_Employed', 'Loan_Amount_Term', 'Credit_History']
for col in null_col :
    null_dict[col] = loan_data[col].mode()[0]
    
null_dict['LoanAmount'] = loan_data['LoanAmount'].median()

null_dict

with open('null_dict.pkl', 'wb') as f :
    pickle.dump(null_dict, f)

with open('null_dict.pkl', 'rb') as f :
    null_dict = pickle.load(f)

----

#2

obj_col = ['Gender', 'Married', 'Dependents', 'Education', 'Self_Employed', 'Property_Area']

#
for col in obj_col :
    label = LabelEncoder()
    loan_data[col] = label.fit_transform(loan_data[col])
    filename = 'label_' + col + '.pkl'
    with open(filename, 'wb') as f :
        pickle.dump(label, f)

----

#최종 모델 학습

import pandas as pd
import numpy as np
import warnings
warnings.filterwarnings('ignore')
import pickle

from sklearn.svm import SVC #SVC 모델을 부르기 위해 필요한 함수
from sklearn.model_selection import train_test_split #모델을 train과 test로
from sklearn.preprocessing import LabelEncoder #문자열->숫자로 변환해주기 위해 필요한 함수
from sklearn.preprocessing import MinMaxScaler #scaler 해주기 위해 필요한 함수

loan_data = pd.read_csv('loan.csv')

#123-1) Real system에 필요한 전처리 코드1)
null_dict = {}

null_col = ['Gender', 'Married', 'Dependents', 'Self_Employed']
for col in null_col :
    null_dict[col] = loan_data[col].mode()[0]
    
null_dict['LoanAmount'] = loan_data['LoanAmount'].median()
null_dict['Loan_Amount_Term'] = loan_data['Loan_Amount_Term'].mode()[0]
null_dict['Credit_History'] = loan_data['Credit_History'].mode()[0]

#123-2)
null_col = ['Gender', 'Married', 'Dependents', 'Self_Employed']
for col in null_col :
    loan_data[col].fillna(loan_data[col].mode()[0], inplace=True)
    
loan_data['LoanAmount'].fillna(loan_data['LoanAmount'].mode()[0], inplace = True)
loan_data['Loan_Amount_Term'].fillna(loan_data['Loan_Amount_Term'].mode()[0], inplace = True)
loan_data['Credit_History'].fillna(loan_data['Credit_History'].mode()[0], inplace = True)

#123-1)과 #123-2)를 합침
null_dict = {}
null_col = ['Gender', 'Married', 'Dependents', 'Self_Employed']
for col in null_col :
    loan_data[col].fillna(loan_data[col].mode()[0], inplace=True)
    null_dict[col] = loan_data[col].mode()[0]
    
null_dict['LoanAmount'] = loan_data['LoanAmount'].median()
null_dict['Loan_Amount_Term'] = loan_data['Loan_Amount_Term'].mode()[0]
null_dict['Credit_History'] = loan_data['Credit_History'].mode()[0]

loan_data['LoanAmount'].fillna(loan_data['LoanAmount'].mode()[0], inplace = True)
loan_data['Loan_Amount_Term'].fillna(loan_data['Loan_Amount_Term'].mode()[0], inplace = True)
loan_data['Credit_History'].fillna(loan_data['Credit_History'].mode()[0], inplace = True)

obj_col = ['Gender', 'Married', 'Dependents', 'Education', 'Self_Employed', 'Property_Area']

# 모델을 학습할 때 필요한 코드 및 Realsystem에 필요한 전처리 코드2)
for col in obj_col :
    label = LabelEncoder()
    loan_data[col] = label.fit_transform(loan_data[col])
    filename = 'label_' + col + '.pkl'
   # with open(filename, 'wb') as f : #real_system에서 불러오기 위해서 파일을 저장해야함. 따라서, real 
   #     pickle.dump(label, f)        #

scaler = MinMaxScaler()

X = loan_data.drop(['Loan_ID', 'Loan_Status'], axis = 1)
Y = loan_data['Loan_Status']

x_train, x_test, y_train, y_test = train_test_split(X, Y, test_size=0.2, stratify=Y, random_state=0)

x_train = scaler.fit_transform(x_train)

x_test = scaler.transform(x_test)

model2 = SVC()

model2.fit(x_train, y_train)

model2.score(x_train, y_train)

model2.score(x_test, y_test)

#최종 모델 저장

with open('model2.pkl', 'wb') as f :
    pickle.dump(model2, f)

---

*머신러닝 순서
1)null_dict.pkl(필요 없는 값들의 각각의 칼럼 위치에 적절한 값(최빈값 또는 중앙값)으로 대체해주기 위해서 이 작업을 한다.)
2)6가지(filename = 'label_' + col + '.pkl')(문자열->숫자로 치환, 왜? 머신러닝은 컴퓨터가 계산해주는데 그러기 위해서는 문자열로 넣어주면 컴퓨터가 인식을 못함.
                                           그렇기 때문에 문자를 숫자로 변환시켜줘서 컴퓨터가 머신러닝을 할 수 있게 해주어야 한다.)
3)scaler.pkl(스케일(모델 정확도를 높이기 위함)을 해준다.)
4)model2.pkl(학습시킬 모델을 저장한다.)


순서:
0)함수 및 데이터 불러온다.
1)모델을 학습시킬 때 필요한 전처리 -> 모델 학습 -> 모델 저장(model2.pkl)
2)Real system에 필요한 전처리 2가지(null_dict, obj_col)한 후 저장(null_dict.pkl, 6가지(filename = 'label_' + col + '.pkl'))
3)
