==> Machine Learning/1-1.real sonar_221109(Wed)-checkpoint.ipynb <==
# 돌인지 마인인지 구분하는 시스템 만들기

import pandas as pd
import pickle

real_data = pd.read_csv('sonar_csv.csv')

real_data.drop('Class', axis = 1, inplace = True)

real = real_data.iloc[100].values.reshape(1, -1)

real

with open('scaler.pkl', 'rb') as f :
    scaler_rb = pickle.load(f)

with open('model3.pkl', 'rb') as f :
    pickle_rb = pickle.load(f)

real1 = scaler_rb.transform(real)

pickle_rb.predict(real1)

pickle_rb.predict(real1)[0]

predict = pickle_rb.predict(real1)[0]

if predict == 'Mine' :
    print('지뢰가 나타났습니다. 돔황챠')
else :
    print('그냥 돌입니다.')

print('첫째만 잘하면 되겠다...')

print('ㅇㅈ')
