==> Machine Learning/2.Decision Tree_221115(Tue).ipynb <==
import math
from collections import Counter, defaultdict
from functools import partial

inputs = [
        ({'level':'Senior','lang':'Java','tweets':'no','phd':'no'},   False),
        ({'level':'Senior','lang':'Java','tweets':'no','phd':'yes'},  False),
        ({'level':'Mid','lang':'Python','tweets':'no','phd':'no'},     True),
        ({'level':'Junior','lang':'Python','tweets':'no','phd':'no'},  True),
        ({'level':'Junior','lang':'R','tweets':'yes','phd':'no'},      True),
        ({'level':'Junior','lang':'R','tweets':'yes','phd':'yes'},    False),
        ({'level':'Mid','lang':'R','tweets':'yes','phd':'yes'},        True),
        ({'level':'Senior','lang':'Python','tweets':'no','phd':'no'}, False),
        ({'level':'Senior','lang':'R','tweets':'yes','phd':'no'},      True),
        ({'level':'Junior','lang':'Python','tweets':'yes','phd':'no'}, True),
        ({'level':'Senior','lang':'Python','tweets':'yes','phd':'yes'},True),
        ({'level':'Mid','lang':'Python','tweets':'no','phd':'yes'},    True),
        ({'level':'Mid','lang':'Java','tweets':'yes','phd':'no'},      True),
        ({'level':'Junior','lang':'Python','tweets':'no','phd':'yes'},False)
    ]

inputs[0]

inputs[0][0]

inputs[0][1]

for i in inputs :
    print(i)

for i in inputs :
    print(i[0]['level'])

for i in inputs :
    print(i[0]['level'] == 'Senior')

for data in inputs :
    if data[0]['level'] == 'Senior' :
        print(data)
        
# [data for data in inputs if data[0]['level'] == 'Senior'] 코드가 같다.

[data for data in inputs if data[0]['level'] == 'Senior']

# 오후 수업

import math

n = 10 / T 5 F 5

p = 5 / 10
q = 5 / 10

# 엔트로피 공식
-p * math.log(p, 2) - q * math.log(q, 2)

# 엔트로피 공식

class_probabilities = [5 / 10, 5 / 10] # class_probabilities = [p, q]

def entropy(class_probabilities) :
    return sum(-p * math.log(p, 2) for p in class_probabilities if p) # if p를 해주는 이유는 None값이 들어갔을때 오류가 나기 때문

entropy(class_probabilities)

labeled_data = [data for data in inputs if data[0]['level'] == 'Senior']

labels = [label for _, label in labeled_data] #

total = len(labels)

Counter(labels)

Counter(labels).values()

[count / total for count in Counter(labels).values()] #True/False의 확률 구하는 공식

#
def class_probabilities(labels) : #
    total = len(labels)
    return [count / total for count in Counter(labels).values()]

#
def data_entropy(labeled_data) : # (labeled_data)를 data_entropy라는 함수에 넣어라.
    labels = [label for _, label in labeled_data] # 언더바 되어 있는 부분은 출력하지말고, label에 해당하는 값만 출력해라.
    probabilities = class_probabilities(labels) # (labels)를 
    return entropy(probabilities) #

labeled_data = [data for data in inputs if data[0]['level'] == 'Senior']

labeled_data = [data for data in inputs if data[0]['level'] == 'Mid']

labeled_data = [data for data in inputs if data[0]['level'] == 'Junior']

data_entropy(labeled_data)

# defaultdict *****
temp = defaultdict(list) # defaultdict를 사용하면 ()안에 있는 형태로 비어있는 value를 만들 수 있음

temp['a']

test = {}

test['Senior'] = ['a']

test['Senior'].append('b')

test

groups = defaultdict(list)
for i in inputs :
    key = i[0]['level']
    print(key)

groups = defaultdict(list)
for i in inputs :
    key = i[0]['level']
    groups[key].append(i)

groups

#
def partition_by(inputs, attribute) :
    groups = defaultdict(list)
    for i in inputs :
        key = i[0][attribute]
        groups[key].append(i)
    return groups

for key in ['level', 'lang', 'tweets', 'phd'] :
    print(key, partition_by(inputs, key))
    print('---------------------------')

for key in ['level', 'lang', 'tweets', 'phd'] :
    print(key, partition_by(inputs, key).keys())
    print('---------------------------')

for key in ['level', 'lang', 'tweets', 'phd'] :
    print(key, partition_by(inputs, key).values())
    print('---------------------------')

for i in partition_by(inputs, 'level').values() :
    print(i)

subsets = partition_by(inputs, 'level').values()

total_count = sum(len(subset) for subset in subsets)

total_count

sum(len(subset) / total_count * data_entropy(subset) for subset in subsets)

# partition별 엔트로피를 구할 것이다.
def partition_entropy(subsets) :
    total_count = sum(len(subset) for subset in subsets)
    return sum(len(subset) / total_count * data_entropy(subset) for subset in subsets)
#엔트로피의 평균을 구해서 ~
# 𝑃   𝑆𝑒𝑛𝑖𝑜𝑟     ∗ 𝐸𝑛𝑡𝑟𝑜𝑝𝑦   2,3     + 𝑃   𝑀𝑖𝑑     ∗ 𝐸𝑛𝑡𝑟𝑜𝑝𝑦   4,0     + 𝑃   𝐽𝑢𝑛𝑖𝑜𝑟       ∗ 𝐸𝑛𝑡𝑟𝑜𝑝𝑦   3,2     =  0.694

def partition_entropy_by(inputs, attribute) :
    partitions = partition_by(inputs, attribute)
    return partition_entropy(partitions.values())

for key in ['level', 'lang', 'tweets', 'phd'] :
    print(key, partition_entropy_by(inputs, key))

mid_inputs = [(data, label) for data, label in inputs if data['level'] == 'Mid'] # data의 'level'인 'Mid'인 녀석들만 뽑아와라.

for key in ['lang', 'tweets', 'phd'] :
    print(key, partition_entropy_by(mid_inputs, key))

senior_inputs = [(data, label) for data, label in inputs if data['level'] == 'Senior'] # data의 'level'인 'Senior'인 녀석들만 뽑아와라.

for key in ['lang', 'tweets', 'phd'] :
    print(key, partition_entropy_by(senior_inputs, key))

junior_inputs = [(data, label) for data, label in inputs if data['level'] == 'Junior'] # data의 'level'인 'Senior'인 녀석들만 뽑아와라.

for key in ['lang', 'tweets', 'phd'] :
    print(key, partition_entropy_by(junior_inputs, key))

#재귀 함수

#재귀 함수 예시
def func_test(x) :
    if x < 0 :
        return
    x = x - 1
    print(x)
    return func_test(x)

func_test(10)

#재귀 함수 예시
def func_test(x) :
    if x < 0 :
        return 0
    x = x - 1
    print(x)
    x = func_test(x)
    print(x)
    return x

func_test(10)

def build_tree_id3(inputs, split_candidates = None) :
    if split_candidates is None :
        split_candidates = inputs[0][0].keys()
    print(split_candidates)

inputs[0]

inputs[0][0]

inputs[0][0].keys()

#
def build_tree_id3(inputs, split_candidates = None) :
    if split_candidates is None :
        split_candidates = inputs[0][0].keys()
        
    num_inputs = len(inputs) # inputs의 전체 갯수는 14개
    num_trues = len([label for _, label in inputs if label])
    print(num_trues)

build_tree_id3(inputs) # num_trues의 갯수

#
def build_tree_id3(inputs, split_candidates = None) :
    if split_candidates is None :
        split_candidates = inputs[0][0].keys()
        
    num_inputs = len(inputs) # inputs의 전체 갯수는 14개
    num_trues = len([label for _, label in inputs if label])
    num_falses = num_inputs - num_trues
    
    print(num_falses)

build_tree_id3(inputs) # num_falses 의 갯수

test = ['abc12321', 'abc21324', 'abcde0', 'abcdef5']

def back(x) :
    return x[-1]

min(test, key = back) #

max(test, key = back)

def back(x, y) :
    return int(x[-1]) + y

min(test, key = partial(back, y = 1)) #

#
def build_tree_id3(inputs, split_candidates = None) :
    if split_candidates is None :
        split_candidates = inputs[0][0].keys()
        # ['level', 'lang', 'tweets', 'phd']
        
    num_inputs = len(inputs) # inputs의 전체 갯수는 14개
    num_trues = len([label for _, label in inputs if label])
    num_falses = num_inputs - num_trues
    
    best_attribute = min(split_candidates, key = partial(partition_entropy_by, inputs)) #의미: ['level', 'lang', 'tweets', 'phd']중에서 엔트로피가 가장 낮은 값을 best_attribute에 넣어서 출력하면 'level'이 나오는 것을 알 수 있음
    
    print(best_attribute)

build_tree_id3(inputs)

#
def build_tree_id3(inputs, split_candidates = None) :
    if split_candidates is None :
        split_candidates = inputs[0][0].keys()
        # ['level', 'lang', 'tweets', 'phd']
        
    num_inputs = len(inputs) # inputs의 전체 갯수는 14개
    num_trues = len([label for _, label in inputs if label])
    num_falses = num_inputs - num_trues
    
    best_attribute = min(split_candidates, key = partial(partition_entropy_by, inputs)) #의미: ['level', 'lang', 'tweets', 'phd']중에서 엔트로피가 가장 낮은 값을 best_attribute에 넣어서 출력하면 'level'이 나오는 것을 알 수 있음
    
    partitions = partition_by(inputs, best_attribute)
    
    print(partitions)

build_tree_id3(inputs)

#
def build_tree_id3(inputs, split_candidates = None) :
    if split_candidates is None :
        split_candidates = inputs[0][0].keys()
        # ['level', 'lang', 'tweets', 'phd']
        
    num_inputs = len(inputs) # inputs의 전체 갯수는 14개
    num_trues = len([label for _, label in inputs if label])
    num_falses = num_inputs - num_trues
    
    best_attribute = min(split_candidates, key = partial(partition_entropy_by, inputs)) #의미: ['level', 'lang', 'tweets', 'phd']중에서 엔트로피가 가장 낮은 값을 best_attribute에 넣어서 출력하면 'level'이 나오는 것을 알 수 있음
    
    partitions = partition_by(inputs, best_attribute)
    new_candidates = [a for a in split_candidates if a != best_attribute]
    
    print(new_candidates)

build_tree_id3(inputs)

#
def build_tree_id3(inputs, split_candidates = None) :
    if split_candidates is None :
        split_candidates = inputs[0][0].keys()
        # ['level', 'lang', 'tweets', 'phd']
        
    num_inputs = len(inputs) # inputs의 전체 갯수는 14개
    num_trues = len([label for _, label in inputs if label])
    num_falses = num_inputs - num_trues
    
    best_attribute = min(split_candidates, key = partial(partition_entropy_by, inputs)) #의미: ['level', 'lang', 'tweets', 'phd']중에서 엔트로피가 가장 낮은 값을 best_attribute에 넣어서 출력하면 'level'이 나오는 것을 알 수 있음
    
    partitions = partition_by(inputs, best_attribute)
    new_candidates = [a for a in split_candidates if a != best_attribute]
    
    subtrees = {attribute_value : build_tree_id3(subset, new_candidates) for attribute_value, subset in partitions.items()}
    
    print(subtrees)

build_tree_id3(inputs) # 오류가 나는 이유:  if not split_candidates : 
                                             # return num_trues >= num_falses    이 코드가 없어서 오류가 남

#
def build_tree_id3(inputs, split_candidates = None) :
    if split_candidates is None :
        split_candidates = inputs[0][0].keys()
        # ['level', 'lang', 'tweets', 'phd']
        
    num_inputs = len(inputs) # inputs의 전체 갯수는 14개
    num_trues = len([label for _, label in inputs if label])
    num_falses = num_inputs - num_trues
    
    if not split_candidates :
        return num_trues >= nuㄴm_falses # 이 값이 없어서 오류가 남
    
    best_attribute = min(split_candidates, key = partial(partition_entropy_by, inputs)) #의미: ['level', 'lang', 'tweets', 'phd']중에서 엔트로피가 가장 낮은 값을 best_attribute에 넣어서 출력하면 'level'이 나오는 것을 알 수 있음
    
    partitions = partition_by(inputs, best_attribute)
    new_candidates = [a for a in split_candidates if a != best_attribute]
    
    subtrees = {attribute_value : build_tree_id3(subset, new_candidates) for attribute_value, subset in partitions.items()}
    
    print(subtrees)

build_tree_id3(inputs)

#
def build_tree_id3(inputs, split_candidates = None) :
    if split_candidates is None :
        split_candidates = inputs[0][0].keys()
        # ['level', 'lang', 'tweets', 'phd']
        
    num_inputs = len(inputs) # inputs의 전체 갯수는 14개
    num_trues = len([label for _, label in inputs if label])
    num_falses = num_inputs - num_trues
    
    if not split_candidates :
        return num_trues >= num_falses # 이 값이 없어서 오류가 남
    
    best_attribute = min(split_candidates, key = partial(partition_entropy_by, inputs)) #의미: ['level', 'lang', 'tweets', 'phd']중에서 엔트로피가 가장 낮은 값을 best_attribute에 넣어서 출력하면 'level'이 나오는 것을 알 수 있음
    
    partitions = partition_by(inputs, best_attribute)
    new_candidates = [a for a in split_candidates if a != best_attribute]
    
    subtrees = {attribute_value : build_tree_id3(subset, new_candidates) for attribute_value, subset in partitions.items()}
    
    subtrees[None] = num_trues > num_falses
    return best_attribute, subtrees

build_tree_id3(inputs)

def classify(tree, x) :
    if tree in [True, False] :
        return tree
    
    attribute, subtree_dict = tree
    # lever, {}
    
    subtree_key = x[attribute]
    # Junior
    
    if subtree_key not in subtree_dict :
        subtree_key = None
        
    subtree = subtree_dict[subtree_key]
    
    
    return classify(subtree, x)

tree = build_tree_id3(inputs)

inputs[0]

classify(tree, {'level': 'Senior', 'lang': 'Java', 'tweets': 'no', 'phd': 'no'})

classify(tree, {'level': 'Junior', 'lang': 'Java', 'tweets': 'yes', 'phd': 'no'})

---

# 최종 정리

import math
from collections import Counter
from functools import partial

#data set
inputs = [
        ({'level':'Senior','lang':'Java','tweets':'no','phd':'no'},   False),
        ({'level':'Senior','lang':'Java','tweets':'no','phd':'yes'},  False),
        ({'level':'Mid','lang':'Python','tweets':'no','phd':'no'},     True),
        ({'level':'Junior','lang':'Python','tweets':'no','phd':'no'},  True),
        ({'level':'Junior','lang':'R','tweets':'yes','phd':'no'},      True),
        ({'level':'Junior','lang':'R','tweets':'yes','phd':'yes'},    False),
        ({'level':'Mid','lang':'R','tweets':'yes','phd':'yes'},        True),
        ({'level':'Senior','lang':'Python','tweets':'no','phd':'no'}, False),
        ({'level':'Senior','lang':'R','tweets':'yes','phd':'no'},      True),
        ({'level':'Junior','lang':'Python','tweets':'yes','phd':'no'}, True),
        ({'level':'Senior','lang':'Python','tweets':'yes','phd':'yes'},True),
        ({'level':'Mid','lang':'Python','tweets':'no','phd':'yes'},    True),
        ({'level':'Mid','lang':'Java','tweets':'yes','phd':'no'},      True),
        ({'level':'Junior','lang':'Python','tweets':'no','phd':'yes'},False)
    ]

# 엔트로피를 구하는 공식
def entropy(class_probabilities) :
    return sum(-p * math.log(p, 2) for p in class_probabilities if p) # if p를 해주는 이유는 None값이 들어갔을때 오류가 나기 때문

# 엔트로피를 적용해서 각각의 확률 구하기(?)
def class_probabilities(labels) : #
    total = len(labels)
    return [count / total for count in Counter(labels).values()]

# 
def data_entropy(labeled_data) : # (labeled_data)를 data_entropy라는 함수에 넣어라.
    labels = [label for _, label in labeled_data] # 언더바 되어 있는 부분은 출력하지말고, label에 해당하는 값만 출력해라.
    probabilities = class_probabilities(labels) # (labels)를 
    return entropy(probabilities) #

labeled_data = [data for data in inputs if data[0]['level'] == 'Senior']

data_entropy(labeled_data)

labeled_data = [data for data in inputs if data[0]['level'] == 'Mid']

data_entropy(labeled_data)

labeled_data = [data for data in inputs if data[0]['level'] == 'Junior']

data_entropy(labeled_data)

#
def partition_by(inputs, attribute) :
    groups = defaultdict(list)
    for i in inputs :
        key = i[0][attribute]
        groups[key].append(i)
    return groups

# partition별 엔트로피를 구할 것이다.
def partition_entropy(subsets) :
    total_count = sum(len(subset) for subset in subsets)
    return sum(len(subset) / total_count * data_entropy(subset) for subset in subsets) # 𝑃   𝑆𝑒𝑛𝑖𝑜𝑟     ∗ 𝐸𝑛𝑡𝑟𝑜𝑝𝑦   2,3     + 𝑃   𝑀𝑖𝑑     ∗ 𝐸𝑛𝑡𝑟𝑜𝑝𝑦   4,0     + 𝑃   𝐽𝑢𝑛𝑖𝑜𝑟       ∗ 𝐸𝑛𝑡𝑟𝑜𝑝𝑦   3,2     =  0.694
#엔트로피의 평균을 구해서 ~


#
def partition_entropy_by(inputs, attribute) :
    partitions = partition_by(inputs, attribute)
    return partition_entropy(partitions.values())

# 
for key in ['level', 'lang', 'tweets', 'phd'] :
    print(key, partition_entropy_by(inputs, key))

mid_inputs = [(data, label) for data, label in inputs if data['level'] == 'Mid'] # data의 'level'인 'Mid'인 녀석들만 뽑아와라.

# Decision Tree (재귀 함수도 적용 되었다.) / mode.fit()와 관련 
def build_tree_id3(inputs, split_candidates = None) :
    if split_candidates is None :
        split_candidates = inputs[0][0].keys()
        # ['level', 'lang', 'tweets', 'phd']
        
    num_inputs = len(inputs) # inputs의 전체 갯수는 14개
    num_trues = len([label for _, label in inputs if label]) #True의 값들
    num_falses = num_inputs - num_trues  #False의 값들
    
    if not split_candidates :
        return num_trues >= num_falses # 이 값이 없어서 오류가 남
    
    best_attribute = min(split_candidates, key = partial(partition_entropy_by, inputs)) #의미: ['level', 'lang', 'tweets', 'phd']중에서 엔트로피가 가장 낮은 값을 best_attribute에 넣어서 출력하면 'level'이 나오는 것을 알 수 있음
    
    # {'Senior' : [(),(),()], 'Junior' : [(),(),()], 'Mid' : [(),(),()]}
    # {'No' : [(),(),()], 'Yes':[(),(),()]}
    # {'Java' : [(),(),()], 'Python': [(),(),()]}
    #{'No': [()], 'Yes' : [()]}
    partitions = partition_by(inputs, best_attribute)
    new_candidates = [a for a in split_candidates if a != best_attribute]
    
    subtrees = {attribute_value : build_tree_id3(subset, new_candidates) for attribute_value, subset in partitions.items()}
    
    subtrees[None] = num_trues > num_falses
    return best_attribute, subtrees

#classify / mode.predict()와 관련
def classify(tree, x) :
    if tree in [True, False] :
        return tree
    
    attribute, subtree_dict = tree
    # lever, {}
    
    subtree_key = x[attribute]
    # Junior
    
    if subtree_key not in subtree_dict :
        subtree_key = None
        
    subtree = subtree_dict[subtree_key]
    
    
    return classify(subtree, x)

from sklearn.tree import DecisionTreeClassifier

model = DecisionTreeClassifier()

model.fit(x_train, y_train)
mode.predict()

---

for _, label in labeled_data :
    print(label)

for data, label in labeled_data :
    print(data, '/', label)

for _, label in labeled_data : # _: 해당하는 변수를 출력하지 않겠다. / 위에 코드랑 같음
    print(label)

for _, label in labeled_data

---
