==> Machine Learning/2.eleveinspect_221103(목)-checkpoint.ipynb <==
# 데이터를 저장한 후 데이터를 불러와서 ~


# 2번째거 활용 (openapi/service/ElevatorInformationService/getElevatorView)

# lastInspctDe(검사일자) / inspctinstt(검사기관) / lastResultNm(합격유무)

#데이터 프레임에 추가하기

import requests
from bs4 import BeautifulSoup
import pandas as pd

import warnings
warnings.filterwarnings('ignore')

df = pd.read_csv('elevInfo.csv', encoding='euc-kr')

df

data = []

url = 'http://apis.data.go.kr/openapi/service/ElevatorInformationService/getElevatorView'

serviceKey = 'BAc5c%2B7CgdSM2dp9F0gq%2FN7FR4TwTghq%2BzHtk7BmmX%2B588SuKYRFAW77xC6eaS6kpEZgWxZnz9a1TSk%2FO%2F6gUg%3D%3D'
elevator_no = '0081201' 

params = {'serviceKey' : serviceKey, 'elevator_no' : elevator_no}
req = requests.get(url, params=params,verify=False)
soup = BeautifulSoup(req.text,'lxml')
col_list = ['lastInspctDe', 'inspctinstt', 'lastResultNm']



for no in df['elevatorNo'] :
    print(str(no).zfill(7))

url = 'http://apis.data.go.kr/openapi/service/ElevatorInformationService/getElevatorView'
serviceKey = 'BAc5c%2B7CgdSM2dp9F0gq%2FN7FR4TwTghq%2BzHtk7BmmX%2B588SuKYRFAW77xC6eaS6kpEZgWxZnz9a1TSk%2FO%2F6gUg%3D%3D'
elevator_no = '0081201' 


params = {'serviceKey' : serviceKey, 'elevator_no' : elevator_no}

req = requests.get(url, params=params)

# 데이터 불러와서 저장해야하는데 안불러와졌음 ㅠㅠㅠ

from tqdm import tqdm

data = []
for no in tqdm(df['elevatorNo']) :
    url = 'http://apis.data.go.kr/openapi/service/ElevatorInformationService/getElevatorView'
    serviceKey = 'BAc5c+7CgdSM2dp9F0gq/N7FR4TwTghq+zHtk7BmmX+588SuKYRFAW77xC6eaS6kpEZgWxZnz9a1TSk/O/6gUg=='
    elevator_no = str(no).zfill(7)

    params = {'serviceKey' : serviceKey, 'elevator_no' : elevator_no}
    req = requests.get(url, params=params, verify=False)
    soup = BeautifulSoup(req.text,'xml')
    
    temp = []
    
    temp.append(no)
    try : 
        temp.append(soup.select('lastInspctDe')[0].text)
    except :
        temp.append('')
    
    try :
        temp.append(soup.select('inspctinstt')[0].text)
    except :
        temp.append('')
        
    try :
        temp.append(soup.select('lastResultNm')[0].text)
    except :
        temp.append('')
        
    data.append(temp)

# 세은 코드

data = []
for no in tqdm(df['elevatorNo']) :
    url = 'http://apis.data.go.kr/openapi/service/ElevatorInformationService/getElevatorView'
    serviceKey = '0kYXnKBN3P1BPz+PbMj53+Xf9Lasd9mZvYohzWDq+10Tagyqazw0pf62r0VMYgKLZZednlK4k5uPA6V49OvEcw=='
    elevator_no = str(no).zfill(7)
    params = {'serviceKey' : serviceKey, 'elevator_no' : elevator_no}
    req = requests.get(url, params=params,verify=False)
    soup = BeautifulSoup(req.text, 'xml')
    temp = []
    temp.append(no)
    try :
        temp.append(soup.select('lastInspctDe')[0].text)
    except :
        temp.append('')
    try :
        temp.append(soup.select('inspctInstt')[0].text)
    except :
        temp.append('')
    try :
        temp.append(soup.select('lastResultNm')[0].text)
    except :
        temp.append('')
    data.append(temp)

data = []
for no in tqdm(df['elevatorNo']) :
    url = "http://apis.data.go.kr/openapi/service/ElevatorInformationService/getElevatorView"
    serviceKey = "tyyr+B6qsLBQB+qXcI9AT0xuf/lSA4MYtsc+r2fj7c9BPR3RgG3P20Yc3Ky6/lW/kJJptIKlrEz/yEVJidDTcg=="
    elevator_no = str(no).zfill(7)
    
    req = requests.get(url,params=params,verify=False)
    soup = BeautifulSoup(req.text,'lxml')
    col_list = ['lastInspctDe','inspctInstt','lastResultNm']
    
    temp = []
    temp.append(no)
    temp.append(soup.select('lastInspctDe')[0].text)
    temp.append(soup.select('inspctInstt')[0].text)
    temp.append(soup.select('lastResultNm')[0].text)
    data.append(temp)

data = []

url = 'http://apis.data.go.kr/openapi/service/ElevatorInformationService/getElevatorView'

http://apis.data.go.kr/openapi/service/ElevatorInformationService/getElevatorView
?serviceKey=BAc5c%2B7CgdSM2dp9F0gq%2FN7FR4TwTghq%2BzHtk7BmmX%2B588SuKYRFAW77xC6eaS6kpEZgWxZnz9a1TSk%2FO%2F6gUg%3D%3D
&elevator_no=0081201

serviceKey = 'BAc5c%2B7CgdSM2dp9F0gq%2FN7FR4TwTghq%2BzHtk7BmmX%2B588SuKYRFAW77xC6eaS6kpEZgWxZnz9a1TSk%2FO%2F6gUg%3D%3D'
elevator_no = '0081201' 

params = {'serviceKey' : serviceKey, 'elevator_no' : elevator_no}
req = requests.get(url, params=params,verify=False)
soup = BeautifulSoup(req.text,'lxml')
col_list = ['lastInspctDe', 'inspctinstt', 'lastResultNm']



for item in soup.items :
    temp = []
    for col in col_list :
        try :
            temp.append(item.select(col)[0].text)
        except :
            temp.append('')
    data.append(temp)

for no in df['elevatorNo'] :
    print(str(no).zfill(7))

http://apis.data.go.kr/openapi/service/ElevatorInformationService/getElevatorView?serviceKey=BAc5c%2B7CgdSM2dp9F0gq%2FN7FR4TwTghq%2BzHtk7BmmX%2B588SuKYRFAW77xC6eaS6kpEZgWxZnz9a1TSk%2FO%2F6gUg%3D%3D&elevator_no=0081201

df2 = pd.read_csv('elevInfo.csv', encoding = 'euckr')

df2

df_all = pd.merge(df, df2, on='elevatorNo') #df1과 df2를 합쳐라. 'elevatorNo'을 기준으로

df_all

-----------------------------------------------------------------------------------------------------------
