==> Machine Learning/2.heart_221109(Wed)-checkpoint.ipynb <==
# Heart Disease Data Set
## 심장 질환 여부 판별하기('target': label)

import pandas as pd
from sklearn.linear_model import LogisticRegression #Classification의 한 종류
from sklearn.model_selection import train_test_split

heart = pd.read_csv('heart.csv')

heart

heart['sex'].value_counts() #남자: 1, 여자: 0 (남자가 여자보다 많다.)

heart['target'].value_counts()

heart.info()

heart.describe()

heart.isnull().sum()

heart['target']

heart['target'].unique()

obj_col = ['target']

for col in obj_col :
    print(col, heart[col].unique())

for col in obj_col :
    print(col, '\n', heart[col].value_counts())

from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_absolute_error

x = heart.drop('target', axis = 1)

y = heart['target']

x_train, x_test, y_train, y_test = train_test_split(x, y, test_size = 0.2, stratify = y, random_state = 0) # stratify =  -> 기준점

print(x_train.shape, y_train.shape) #학습시킬 문제, 학습시킬 문제에 대한 정답
print(x_test.shape, y_test.shape) #시험문제, 시험문제에 대한 정답

model = LogisticRegression() #모델 선언

# train

model.fit(x_train, y_train) # 공부(학습)시킬 때 필요한 함수

pre_train = model.predict(x_train) #트레인에 대한 예측값, pre_train

mean_absolute_error(pre_train, y_train)

model.score(x_train, y_train)

# test

pre_test = model.predict(x_test) #트레인에 대한 예측값, pre_train

mean_absolute_error(pre_test, y_test)

model.score(x_test, y_test)

----------------------

# scaler 적용 했을 때

import pandas as pd
from sklearn.linear_model import LogisticRegression #Classification의 한 종류
from sklearn.model_selection import train_test_split



heart = pd.read_csv('heart.csv')

heart

for a in heart.drop('target', axis=1).columns :
    print(a)
    print(heart[a].mean())
    print(heart[a].median())
    print('---------------')

x = heart.drop('target', axis = 1)

y = heart['target']

x_train, x_test, y_train, y_test = train_test_split(x, y, test_size = 0.2, random_state = 0) # stratify =  -> 기준점

print(x_train.shape, y_train.shape) #학습시킬 문제, 학습시킬 문제에 대한 정답
print(x_test.shape, y_test.shape) #시험문제, 시험문제에 대한 정답

from sklearn.preprocessing import StandardScaler

scaler1 = StandardScaler() # 평균과 표준편차 

x_train = scaler1.fit_transform(x_train)
x_test= scaler1.transform(x_test)
# x_test = scaler.transform(x_test)

x_train

model2 = LogisticRegression() #모델 선언

# train

model2.fit(x_train, y_train) # 공부(학습)시킬 때 필요한 함수

model2.predict(x_train)

print(x_train.shape, y_train.shape)

model2.score(x_train, y_train)

model2.score(x_test, y_test)

real = x.iloc[0].values

real = [35, 1, 2, 135, 240, 0, 0, 160, 1, 2.1, 1, 0, 1]

real

real = x.iloc[0].values.reshape(1, -1)

real

real = scaler1.transform(real)

real

model2.predict(real)

predict1 = model2.predict(real)[0]

print(predict1)

import pickle

with open('scaler1.pkl', 'wb') as f :
    pickle.dump(scaler1, f)

with open('model2.pkl', 'wb') as f :
    pickle.dump(model2, f)

----

#PolynomialFeatures 모델을 적용했을 때

import pandas as pd
import numpy as np
from sklearn.preprocessing import PolynomialFeatures #Classification의 한 종류
from sklearn.model_selection import train_test_split



x = heart.drop('target', axis = 1)
y = heart['target']

from sklearn.preprocessing import PolynomialFeatures

pol = PolynomialFeatures(degree=2)

x = pol.fit_transform(x)

x_train, x_test, y_train, y_test = train_test_split(x, y, test_size = 0.2, random_state = 0) # stratify =  -> 기준점

model4 = PolynomialFeatures()

model4.fit(x_train, y_train) # 공부(학습)시킬 때 필요한 함수

model4.predict(x_train)

x_train

y_train

model4.score(x_train, y_train)

model4.score(x_test, y_test)

import numpy as np

real = [35, 1, 2, 135, 240, 0, 0, 160, 1, 2.1, 1, 0, 1]

real = np.asarray(real).reshape(1, -1) # ???

real = pol.transform(real)

model4.predict(real)

---

(w1x1+w2x2+w3x3+w4x4+w5x5) + b #행렬의 곱 (선형대수학에서의 매트릭스 개념)

2X2 2X3 -> 2X3 (numpy가 행렬의 곱을 잘 계산한다.)

x.columns

model.coef_ # W값

model.intercept_

model.score(x_train, y_train)

----
