==> Machine Learning/2.real loan-checkpoint.ipynb <==
# Real system

import pandas as pd
import numpy as np
import pickle

loan_real = ['LP001345', None, 'Yes', '2', 'Not Graduate', 'No', 4288, 3263.0, 133.0, 180.0, 1.0, 'Urban']

loan_real = np.asarray(loan_real).reshape(1, -1)

loan_df = pd.DataFrame(loan_real, columns = ['Loan_ID', 'Gender', 'Married', 'Dependents', 'Education', 'Self_Employed', 'ApplicantIncome', 'CoapplicantIncome', 'LoanAmount', 'Loan_Amount_Term', 'Credit_History', 'Property_Area'])

loan_df

with open('null_dict.pkl', 'rb') as f :
    null_dict = pickle.load(f)

null_col = ['Gender', 'Married', 'Dependents', 'Self_Employed', 'LoanAmount', 'Loan_Amount_Term', 'Credit_History']
for col in null_col :
    loan_df[col].fillna(null_dict[col], inplace=True)

null_dict

loan_df

obj_col = ['Gender', 'Married', 'Dependents', 'Education', 'Self_Employed', 'Property_Area']

for col in obj_col :
    filename = 'label_' + col + '.pkl'
    #print(filename)
    with open(filename, 'rb') as f :
         label = pickle.load(f)
    loan_df[col] = label.transform(loan_df[col])

loan_df.drop('Loan_ID', axis = 1, inplace = True)

loan_df

with open('scaler.pkl', 'rb') as f :
    scaler = pickle.load(f)

loan_df = scaler.transform(loan_df)

with open('model2.pkl', 'rb') as f :
    model2 = pickle.load(f)

predict = model2.predict(loan_df)

if predict == 'Y' :
    print('대출 승인!!!!')
else :
    print('대출 거절 ㅠㅠ')
