==> Machine Learning/2-1.diabetes_real_221110(Thru)-checkpoint.ipynb <==
import pandas as pd
import numpy as np
import preprocess as pre

pre.col_list

diabetes_data = pd.read_csv('diabetes.csv')

real = [  1.   , 163.   ,  72.   ,   0.   ,   0.   ,  39.   ,   1.222,
        33.   ]

real = np.asarray(real).reshape(1, -1)

real

df = pd.DataFrame(real, columns = ['Pregnancies', 'Glucose', 'BloodPressure', 'SkinThickness', 'Insulin', 'BMI', 'DiabetesPedigreeFunction', 'Age'])

df

# 나이별로 구간 나누기
def ageRange(age) :
    if 21 <= age < 30 :
        return 1
    elif 30 <= age < 40 :
        return 2
    elif 40 <= age < 50 :
        return 3
    elif 50 <= age < 60 :
        return 4
    elif 60 <= age < 70 :
        return 5
    else :
        return 6    

col_list = ['Glucose','BloodPressure','SkinThickness','Insulin','BMI']

for col in col_list :
    df[col] = np.where(df[col] == 0, np.nan, df[col]) #= df[col] = df[col].str.replace(0, np.nan)

df['Age2'] = df['Age'].apply(ageRange)

for col in col_list :
    df[col].fillna(df.groupby('Age2')[col].transform('median'), inplace = True)

import pickle

#
with open('dict_tot.pkl', 'rb') as f :
    dict_tot = pickle.load(f)

df.loc[0, 'SkinThickness'] = dict_tot['SkinThickness'][df.iloc[0]['Age2']]

df

for col in col_list :
    df.loc[0, col] = dict_tot[col][df.loc[0]['Age2']]

df.drop('Age2', axis = 1, inplace = True)

#
with open('dict_tot.pkl', 'rb') as f :
    model = pickle.load(f)

predict = model.predict(df)

if predict == 0 :
    print('당뇨가 없습니다')
else :
    print('당뇨 위험')
