==> Machine Learning/2-1.real heart_221109(Wed)-checkpoint.ipynb <==
import pandas as pd
import pickle

real_data = pd.read_csv('heart.csv')

real_data.drop('target', axis = 1, inplace = True)

real = real_data.iloc[100].values.reshape(1, -1)

real

with open('scaler1.pkl', 'rb') as f :
    scaler1_rb = pickle.load(f)

with open('model2.pkl', 'rb') as f :
    pickle1_rb = pickle.load(f)

real2 = scaler1_rb.transform(real)

pickle1_rb.predict(real2)

pickle1_rb.predict(real2)[0]

predict = pickle1_rb.predict(real2)[0]
