==> Machine Learning/3.titanic2-2_221031(월)-checkpoint.ipynb <==
import pandas as pd
import numpy as np

train = pd.read_csv('titanic/train.csv')

test = pd.read_csv('titanic/test.csv')

print(train.shape, test.shape) #test 데이터 프레임에는 'Survived'의 속성이 없기 때문에 11개임

train_test_dataset = [train, test]

# data.groupby('Title')['Age'].transform('median')값을 원래는 train_title_age라는 변수로 지정해서 표기해줘야한다.
# train_title_age = data.groupby('Title')['Age'].transform('median')
# 값을 고정 안시키면 데이터 값이 바뀔 수 있기 때문에 꼭 변수를 선언해줘서 전처리 과정을 돌려야한다.

for data in train_test_dataset :
    data['Title'] = data['Name'].str.extract('([a-zA-Z]+)\.') 
    data['Title'] = data['Title'].str.lower()
    
    data['Title'] = np.where(data['Title'] == 'mr', 0, 
                           np.where(data['Title'] == 'miss', 1, 
                                    np.where(data['Title'] == 'mrs', 2, 3)))
    
    data['Age'] = data['Age'].fillna(data.groupby('Title')['Age'].transform('median'))
    
    data['Age'] = np.where(data['Age'] <= 16, 0, 
                             np.where(data['Age'] <= 26, 1,
                                      np.where(data['Age'] <= 36, 2,
                                               np.where(data['Age'] <= 62, 3, 4))))
    
    data['Embarked'].fillna('S', inplace=True)
    
    data['Embarked'] = np.where(data['Embarked'] == 'S', 0,
                              np.where(data['Embarked'] == 'C', 1, 2))
    
    data['Cabin'] = data['Cabin'].str[0]
    
    cabin_mapping = {'A': 0.0, 'B': 0.4, 'C': 0.8, 'D': 1.2, 'E': 1.6, 'F': 2.0, 'G': 2.4, 'T': 2.8}
    
    data['Cabin'] = data['Cabin'].map(cabin_mapping)
    data['Cabin'].fillna(data.groupby('Pclass')['Cabin'].transform('median'), inplace=True)
    
    data['Sex'] = data['Sex'].map({'male':0, 'female':1})
    
    data['Fare'].fillna(data.groupby('Pclass')['Fare'].transform('median'), inplace=True)
    data['Fare'] = np.where(data['Fare'] <= 17, 0, 
                          np.where(data['Fare'] <= 30, 1, 
                                   np.where(data['Fare'] <= 100, 2, 3)))
    
    data['FamilySize'] = data['SibSp'] + data['Parch'] + 1
    family_mapping = {1: 0.0, 2: 0.4, 3: 0.8, 4: 1.2, 5: 1.6, 6: 2.0, 7: 2.4, 8: 2.8, 9: 3.2, 10: 3.6, 11: 4.0}
    data['FamilySize'] = data['FamilySize'].map(family_mapping)
    

features_drop = ['PassengerId', 'Name', 'Ticket', 'SibSp', 'Parch']

train = train.drop(features_drop, axis = 1)

test = test.drop(features_drop, axis = 1)

train.head(1)

test.head(1)

# train에 대한 모델링 작업

train.isnull().sum() #학습지, ex)모의고사

test.isnull().sum() #실전, ex)수능

# 이렇게 하는 이유는?

from sklearn.linear_model import LogisticRegression
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import GradientBoostingClassifier
from sklearn.ensemble import AdaBoostClassifier
from sklearn.metrics import accuracy_score
from sklearn.model_selection import train_test_split

train

train.drop('Survived', axis = 1) #train.drop(columns = 'Survived')

inputs = train.drop('Survived', axis = 1) #train.drop(columns = 'Survived')
#문제, X

outputs = train['Survived']
#정답지, Y

print(inputs.shape, outputs.shape) #*****문제와 정답의 갯수가 일치해야 한다. *****

train_input, test_input, train_output, test_output = train_test_split #학습이 잘되었는지 눈으로 확인하기 위해서 이 작업을 한다.

train_input, test_input, train_output, test_output = \
train_test_split(inputs, outputs, test_size = 0.2, stratify = outputs, random_state = 10)
#랜덤 값(random_state)은 고정 값(seed)을 항상 주는게 좋다.
#stratify = outputs 뜻은 train과 test 데이터의 비율을 맞추기 위함.
#순서: 1)train_input, 2)test_input, 3)train_output, 4)test_output

print(train_input.shape, train_output.shape)
print(test_input.shape, test_output.shape)

outputs.value_counts()

train_output.value_counts()

test_output.value_counts()

model1 = LogisticRegression()
model2 = DecisionTreeClassifier()
model3 = RandomForestClassifier()
model4 = GradientBoostingClassifier()
model5 =AdaBoostClassifier()

# 모델1

model1.fit(train_input, train_output) #input은 다 만듬. 머신러닝 모델을 만듬. (패키지를 이용해서 모델을 쉽게 만들 수 있음)

model1.score(train_input, train_output) * 100 #모델1에 대한 학습데이터

model1.score(test_input, test_output) * 100 #

test_pre = model1.predict(test_input)

#예시
test_input.iloc[0]

test_pre[0] #예시에서의 예측값

test_output.iloc[0] #예시에서의 실제값

test_pre

accuracy_score(test_pre, test_output) * 100 #

# 모델2

model2.fit(train_input, train_output)

model2.score(train_input, train_output) * 100

model2.score(test_input, test_output) * 100

test_pre = model2.predict(test_input)

# 모델3

model3.fit(train_input, train_output)

model3.score(train_input, train_output) * 100

model3.score(test_input, test_output) * 100

test_pre = model3.predict(test_input)

# 모델4

model = model4.fit(train_input, train_output)

model4.score(train_input, train_output) * 100

model4.score(test_input, test_output) * 100

test_pre = model4.predict(test_input)

model

import pickle #파이썬에 있는 기본 라이브러리이다.

with open('gbc.pkl', 'wb') as f : #'wb'는 write binary
    pickle.dump(model, f)

여기까지 한 수업내용이 모델을 만든 과정이다.

여기, 'rb' 구간 부터는 저장된 데이터 프레임을 들고 와서 읽는 작업을 하는 구간이다.

with open('gbc.pkl', 'rb') as f : #'rb'는 read binary
    load_model = pickle.load(f)

load_model

test_input

data1 = test_input.iloc[0].to_list()

import numpy as np

np_data = np.asarray(data1)

re_data = np_data.reshape(1, -1)

pred = load_model.predict(re_data)

if pred[0] == 0 :
    print('Died...')
else :
    print('Survived!!!')

# 모델5

model5.fit(train_input, train_output)

model5.score(train_input, train_output) * 100

model5.score(test_input, test_output) * 100

test_pre = model5.predict(test_input)

# 
0.전처리
1.모델 튜닝을 해야한다.
2.모델 학습시키는 곳

model = {'LogisticRegression': LogisticRegression(),
         'DecisionTreeClassifier':DecisionTreeClassifier(),
         'RandomForestClassifier': RandomForestClassifier(),
         'RandomForestClassifier': RandomForestClassifier(n_estimators = 1000, criterion = ""),
         'GradientBoostingClassifier': GradientBoostingClassifier(),
         'AdaBoostClassifier': AdaBoostClassifier() }

#강사님 모델링 방법

model = {'LogisticRegression': LogisticRegression(),
         'DecisionTreeClassifier':DecisionTreeClassifier(),
         'RandomForestClassifier': RandomForestClassifier(),
         'GradientBoostingClassifier': GradientBoostingClassifier(),
         'AdaBoostClassifier': AdaBoostClassifier() }

model_1 = []
train_1 = []
test_1 = []

for m in model :
    clf = model[m]
    clf.fit(train_input, train_output)
    train_score = clf.score(train_input, train_output) * 100
    test_score = clf.score(train_input, train_output) * 100
    
    print((f"[%26s] train:%.2f / test: %.2f") % (m, train_score, test_score))

data = list(zip(model_1, train_1, test_1))

df = pd.DataFrame(data, columns = ['mode1', 'train', 'test'])

df

df.sort_values('train', ascending = False)

# test에 대한 모델링 작업

sub = pd.read_csv('titanic/submission.csv')

answer = sub['Survived']

test #실제 데이터 프레임을 

submission = load_model.predict(test) #예측한 정답은 sub에 채웠다.

submission

accuracy_score(answer, submission)

accuracy_score(answer, submission) * 100

test.describe()

# (참고)
# X AI: 설명 가능한 인공지능

머신러닝이나 AI 등은 실험적인 학문 = 경험적인 학문이다.

sub2 = pd.read_csv('titanic/submission2.csv')

sub2

sub2['Survived'] = submission

sub2

sub2.to_csv('result.csv', index = False)

sub2['Survived'].value_counts()
