==> Machine Learning/3.titanic_221027(목)-checkpoint.ipynb <==
#titanic 머신러닝
#test (문제)
#train ()

import pandas as pd

train = pd.read_csv('titanic/train.csv')

train = pd.read_csv('./titanic/train.csv') #=train = pd.read_csv('titanic/train.csv')

('../ # 상대경로)

train.head(1)

train.info()

train

# PassengerId: 승객 번호 (object)
# Survived : 생존한 사람 (1: 생존한 사람, 0: 사망한 사람) (object)
# Pclass : 등급 (1, 2, 3) (object)
# SibSp: 사촌 + 배우자의 수 (숫자)
# Parch: 부모님 + 아이들 수 (숫자)
# Ticket: 티켓 번호 (object)
# Fare : 요금 (숫자)
# Cabin: 배안의 객실 (object)
# Embarked : 탑승한 선착장 (object)

train.shape #사용할 수 있는 건 11개? 왜?

train.isnull().sum() # .isnull()은 null의 갯수를 파악할 때 사용함

# 분류 (classification): 분류할 때 사용 / 회귀(regression): 값을 예측할 때 사용

# binary classification (2개 분류)
# multi classification (3개 이상 분류)

# 우리가 할 것은 '분류'이다.

# 이 데이터 안에 살아있는 사람과 죽어있는 사람의 수는?

train['Survived'].value_counts() #0: 죽은사람, 1: 생존한 사람

train['Pclass'].value_counts() #0: 죽은사람, 1: 생존한 사람

train['Sex'].value_counts()

남자인데, pclass3에 탄사람은?
성별로 pclass에 탄사람은?
pclass에 타서 생존한 사람은?

#남자인데, pclass3에 탄사람은?
len(train[(train['Sex'] == 'male') & (train['Pclass'] == 3)])

train

len(train[(train['Sex'] == 'female') & (train['Pclass'] == 2)])

# pclass 2 , 생존한사람의 수는?

len(train[(train['Pclass'] == 2) & (train['Survived'] == 1)])

# 강사님 문제

M 0
M 1

M = 50 %

M 0
M 1
M 1

M = 75 %

train.groupby('Sex')['Survived'].mean() #성별이 여자이면 생존 확률 증가

train.groupby('Pclass')['Survived'].mean() #등급이 높을 수록 생존 확률 증가

train.groupby('SibSp')['Survived'].mean().sort_values(ascending=False)

train['SibSp'].value_counts()

train.groupby('Parch')['Survived'].mean().sort_values(ascending=False)

train['Parch'].value_counts()

train.groupby('Embarked')['Survived'].mean()

# pclass별, 성별, 생존 확률은?

train.groupby(['Pclass', 'Sex'])['Survived'].mean().sort_values(ascending=False)

# pcass별, 요금의 평균은?

train.groupby('Pclass')['Fare'].mean()

train.groupby('Embarked')['Fare'].mean()

# pclass별, 성별, survived할 확률 + fare의 평균은?

train.groupby(['Pclass', 'Sex'])[['Survived', 'Fare']].mean().sort_values(by='Fare', ascending=False)

train.groupby(['Pclass', 'Sex'])[['Survived', 'Fare']].mean() #실제로 등급이 높을수록 요금을 많이 썼음

#데이터분석과 머신러닝의 차이점은?

22-10-28(금)은 
-타이타닉 조금 더 한다.
-머신러닝을 하기 전 전처리 배울 예정 
(map과 where)
