==> Machine Learning/3-1.diabet train(self review)_221110-checkpoint.ipynb <==
# Model system

#전처리

import pandas as pd
import numpy as np
import warnings
warnings.filterwarnings('ignore')
import pickle

from sklearn.svm import SVC
from sklearn.model_selection import train_test_split

diabetes_data = pd.read_csv('diabetes.csv')

# 1)
col_list = ['Glucose','BloodPressure','SkinThickness','Insulin','BMI']
for col in col_list :
    diabetes_data[col] = np.where(diabetes_data[col] == 0, np.nan, diabetes_data[col]) #= data[col] = data[col].str.replace(0, np.nan)

# 2) 나이별로 구간 나누기
def ageRange(age) :
    if 21 <= age < 30 :
        return 1
    elif 30 <= age < 40 :
        return 2
    elif 40 <= age < 50 :
        return 3
    elif 50 <= age < 60 :
        return 4
    elif 60 <= age < 70 :
        return 5
    else :
        return 6    

diabetes_data['Age2'] = diabetes_data['Age'].apply(ageRange)

for col in col_list :
    diabetes_data[col].fillna(diabetes_data.groupby('Age2')[col].transform('median'), inplace = True)

#Real system의 전처리를 하기 위해 필요한 코드
dict_tot = {}
for col in col_list :
    index = diabetes_data.groupby('Age2')[col].median().index
    value = diabetes_data.groupby('Age2')[col].median().values
    temp = {}
    for i in range(len(index)) :
        temp[index[i]] = value[i]
    dict_tot[col] = temp

dict_tot

with open('dict_tot.pkl', 'wb') as f :
    pickle.dump(dict_tot, f)

#모델 학습

x = diabetes_data.drop(['Outcome', 'Age2'], axis = 1)

y = diabetes_data['Outcome']

x_train, x_test, y_train, y_test = train_test_split(x, y, test_size = 0.2, random_state = 0)

model = SVC()

model.fit(x_train, y_train)

model.score(x_train, y_train)

model.score(x_test, y_test)

# 모델 저장

with open('model.pkl', 'wb') as f :
    pickle.dump(model, f)

----

diabetes_data.groupby('Age2')['Glucose'].median()
