==> Machine Learning/3-3.diabet preprocess(self review)_221110-checkpoint.ipynb <==
import pandas as pd
import numpy as np
import pickle 
import preprocess as pre

# real 실제 사용자의 데이터
real = [1., 163.,  72., 0., 0.,  39., 1.222, 33.]
real = np.asarray(real).reshape(1,-1)
df = pd.DataFrame(real, columns= ['Pregnancies', 'Glucose', 'BloodPressure', 'SkinThickness', 'Insulin','BMI', 'DiabetesPedigreeFunction', 'Age'])

df

df2 = pre.prefunc(df)

df2

with open('model.pkl','rb') as f :
    model = pickle.load(f) # 머신러닝 뇌  결과 출력 

predict = model.predict(df2)

if predict == 0 :
    print('당뇨가 없습니다.')
else :
    print('당뇨위험')
