==> Machine Learning/5.Basic Pipeline_221130(Wed).ipynb <==
# 오후 수업

# svc 기본 모델로 학습시키기
data load -> standardscaler -> train, test(0.2) -> svc -> fit -> score

from sklearn.datasets import load_wine
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split
from sklearn.svm import SVC

import pandas as pd

wine = load_wine()

wine.feature_names

df = pd.DataFrame(x, columns = wine.feature_names)

df.describe()

# scaler : 값을 정교화(standardize) 시켜주는 작업

x = wine.data # 나머지 변수들

y = wine.target # 종속변수

scaler = StandardScaler()

scaler.fit(x) # 평균과 표준편차 관련을 저장시키기 위해 필요한 작업
scaler.transform(x) # z = () /  공식을 사용하기 위해 이 작업이 필요 (평균과 표준편차 값들이 표준화가 안되어 있기 때문에 값들을 표준화시키기 위해 필요한 작업)

scaler_x = scaler.fit_transform(x)

x_train, x_test, y_train, y_test = train_test_split(scaler_x, y, test_size = 0.2, random_state = 0)

model = SVC()

model.fit(x_train, y_train)

model.score(x_train, y_train)

model.score(x_test, y_test)

---

# svc 기본 모델로 학습시키기
data load -> 전처리.......(타이타닉, loan 데이터셋 등...) -> standardscaler -> train, test(0.2) -> svc -> fit -> score

# Pipeline을 어떻게 쓰나~~~~~~~?

from sklearn.pipeline import Pipeline

pipeline = Pipeline([('scaler2', StandardScaler()), ('svc', SVC())])

x = wine.data # 나머지 변수들

y = wine.target # 종속변수

x_train, x_test, y_train, y_test = train_test_split(x, y, test_size = 0.2, random_state = 0)

pipeline.fit(x_train, y_train)

pipeline.score(x_train, y_train)

pipeline.score(x_test, y_test)

pipeline.named_steps # pipeline 안에 필요한 부분('scaler2' 또는 'svc')을 불러 올 때 사용 

pipeline.named_steps['scaler2'].mean_

pipeline.named_steps['svc']

pipeline.named_steps['svc'].get_params()

from sklearn import set_config

set_config(display = "diagram")

---

from sklearn.pipeline import make_pipeline

pipeline2 = make_pipeline(StandardScaler(), SVC())

pipeline2

pipeline2.fit(x_train, y_train)

pipeline2.score(x_test, y_test)

pipeline2.named_steps

pipeline2.named_steps['standardscaler']

pipeline2.named_steps['standardscaler'].mean_

---

#

from joblib import dump, load

dump(pipeline2, 'pipeline2.joblib')

load_pipeline = load('pipeline2.joblib')

load_pipeline.score(x_test, y_test)

# 개인이 직접 만든 함수로 저장할 수 있나? yes

def func(x):
    print('Before x', x)
    x = x + 1
    print('After x', x)
    return x

from sklearn.preprocessing import FunctionTransformer #

trans_func = FunctionTransformer(func)

pipeline3 = make_pipeline(StandardScaler(), trans_func, SVC())

pipeline3

pipeline3.fit(x_train, y_train)
