==> Machine Learning/6.Real Pipeline_221130(Wed).ipynb <==
# 문제: 중앙값 채우기 / 
null(median) func -> standardscaler -> svc / train, test 0.25
real_data = water.iloc[0].values.reshape(1, -1)


# 강사님 풀이

from sklearn.datasets import load_wine
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split
from sklearn.svm import SVC
from sklearn.pipeline import Pipeline
from sklearn import set_config
from sklearn.pipeline import make_pipeline

import pandas as pd

water = pd.read_csv('drinking_water_potability.csv')

water.columns

water.isnull().sum()

water['ph'].fillna(water['ph'].median(), inplace = True)

# null 값 중앙값으로 채우기
col_list = ['ph', 'Sulfate', 'Trihalomethanes']
for col in col_list :
    water[col].fillna(water[col].median(), inplace = True)

water.isnull().sum()

water[['ph', 'Sulfate', 'Trihalomethanes']].median()

def median_func(data) :
    col_list = ['ph', 'Sulfate', 'Trihalomethanes']
    for col in col_list :
        if col == 'ph' :
            data[col].fillna(7.036752, inplace = True)
        elif col == 'Sulfate' :
            data[col].fillna(333.073546, inplace = True)
        elif col == 'Trihalomethanes':
            data[col].fillna(66.622485, inplace = True)
    
    return data

from sklearn.preprocessing import FunctionTransformer, StandardScaler
from sklearn.pipeline import make_pipeline
from sklearn.model_selection import train_test_split
from sklearn.svm import SVC

medianfunc = FunctionTransformer(median_func)

pipeline = make_pipeline(medianfunc, StandardScaler(), SVC())

X = water.drop('Potability', axis = 1)
Y = water['Potability']

x_train, x_test, y_train, y_test = train_test_split(X, Y, test_size=0.25, random_state=0)

pipeline.fit(x_train, y_train)

pipeline.score(x_test, y_test)

# 설계 끝내고 저장해야함.
from joblib import dump, load

dump(pipeline, 'water_model.joblib')

import pandas as pd

col_names = ['ph', 'Hardness', 'Solids', 'Chloramines', 'Sulfate', 'Conductivity',
       'Organic_carbon', 'Trihalomethanes', 'Turbidity']

real_water = [[None, 2.04890456e+02, 2.07913190e+04, 7.30021187e+00,
        3.68516441e+02, 5.64308654e+02, 1.03797831e+01, 8.69909705e+01,
        2.96313538e+00]]

pd.DataFrame(real_water, columns = col_names)

real_df = pd.DataFrame(real_water, columns = col_names)

from joblib import load

from waterlib import median_func

pipeline = load('water_model.joblib')

pipeline

ret = pipeline.predict(real_df)

pipeline.predict(real_df)

real_df

ret[0]

if ret[0] == 0 :
    print('Do not Drink!!!')
else :
    print('You can Drink')

ph = input('ph를 입력하세요.')

columns = []

# real data 넣기
real_water = [None, 2.04890456e+02, 2.07913190e+04, 7.30021187e+00,
        3.68516441e+02, 5.64308654e+02, 1.03797831e+01, 8.69909705e+01,
        2.96313538e+00]

데이터 프레임으로 컬럼이름 줘서 만들어야겠다.

water

water

water.info()

water.iloc[0].values

water.isnull().sum()

col = ['ph', 'Sulfate', 'Trihalomethanes']

for i in col :
    water[i] = water[i].fillna(water[i].median())

water.isnull().sum()

pipeline = Pipeline([('scaler', StandardScaler()), ('svc', SVC())])

x = water.drop('Potability', axis = 1) # 나머지 변수들

y = water['Potability'] # 종속변수

x_train, x_test, y_train, y_test = train_test_split(x, y, test_size = 0.2, random_state = 0)

pipeline.fit(x_train, y_train)

pipeline.score(x_train, y_train)

pipeline.score(x_test, y_test)

# make_pipeline 활용

pipeline2 = make_pipeline(StandardScaler(), SVC())

pipeline2

pipeline2.fit(x_train, y_train)

pipeline2.score(x_test, y_test)

pipeline2.named_steps

pipeline2.named_steps['standardscaler']

pipeline2.named_steps['standardscaler'].mean_

water_dict = {}
for i in water.columns :
    water_dict[i] = water[i].median()

water_dict

for i in water :
    print(water[i])

def scaler(x) : 
    

scaler(real_water)

trans_func = FunctionTransformer(scaler)

pipeline3 = make_pipeline(tran, SVC())
