==> Machine Learning/8.XGBoost-checkpoint.ipynb <==
pip install xgboost

lstm(딥러닝)

# LightGBM /XGBoost <- Decision Tree 기반 알고리즘

from xgboost import XGBClassifier
from sklearn.ensemble import RandomForestClassifier

from sklearn import datasets

digits = datasets.load_digits() # digits는 이미지 데이터임

from sklearn.model_selection import train_test_split

x_train, x_test, y_train, y_test = train_test_split(digits.data, digits.target, test_size=0.2)

xgb = XGBClassifier()
rf = RandomForestClassifier()

xgb.fit(x_train, y_train)
rf.fit(x_train, y_train)

xgb.score(x_test, y_test)

rf.score(x_test, y_test)
