==> Machine Learning/9.Text-checkpoint.ipynb <==
docs = ["the cat in the hat",
       "the cow jumped over the moon",
       "the cat mooed and the cow meowed", 
       "the cat said to the cow cow you are not a cat"]

# 단어 분리 후 중복된 것을 빼고 싶음

# 강사님 풀이

vocabulary= set(' '.join(docs).split())

common_words = set(['a', 'to', 'the', 'in', 'and', 'are'])

vocabulary

common_words

vocabulary2 = vocabulary - common_words

vocabulary2

#문장마다 검사하기
결과값: 
[{'said': False,
  'you': False,
  'not': False,
  'jumped': False,
  'hat': True,
  'cow': False,
  'over': False,
  'meowed': False,
  'cat': True,
  'moon': False,
  'mooed': False},
 {'said': False,
  'you': False,
  'not': False,
  'jumped': True,
  'hat': False,
  'cow': True,
  'over': True,
  'meowed': False,
  'cat': False,
  'moon': True,
  'mooed': False},
 {'said': False,
  'you': False,
  'not': False,
  'jumped': False,
  'hat': False,
  'cow': True,
  'over': False,
  'meowed': True,
  'cat': True,
  'moon': False,
  'mooed': True},
 {'said': True,
  'you': True,
  'not': True,
  'jumped': False,
  'hat': False,
  'cow': True,
  'over': False,
  'meowed': False,
  'cat': True,
  'moon': False,
  'mooed': False}]

# 강사님 풀이
for d in docs :
    for v in vocabulary2 :
        print(v, v in d)

# 강사님 코드
dic_list = []
for d in docs :
    dic = {}
    for v in vocabulary2 :
        dic[v] = v in d
    dic_list.append(dic)

dic_list

# 강사님 코드 (comprehension 형태)
[{v:(v in d) for v in vocabulary2} for d in docs]

doc_contains = [{v:(v in d) for v in vocabulary2} for d in docs]

import pandas as pd

dcs = pd.DataFrame(doc_contains)

dcs

dcs.sort_index(axis = 1)

# 오후 수업

# 갯수 새는 것을 표기해보자.

dic_list = []
for d in docs :
    dic = {}
    for v in vocabulary2 :
        dic[v] = d.count(v)
    dic_list.append(dic)

dic_list

dcs1 = pd.DataFrame(dic_list)

dcs1.sort_index(axis = 1)

word_count = [{v:(d.count(v)) for v in vocabulary2} for d in docs]

wcs = pd.DataFrame(word_count)

wcs

wcs.sort_index(axis = 1)

from sklearn.feature_extraction.text import CountVectorizer

sparse = CountVectorizer(stop_words='english').fit_transform(docs) #CountVectorizer: 각 칼럼들의 값들을 세주는 것?

sparse

sparse.todense()

# 단어의 빈도수 구하기

dcs

dcs.sum()

doc_freq = pd.DataFrame(dcs.sum(axis=0)).T

#DF(Document Frequency): 특정 단어가 등장한 문서의 수
doc_freq 

doc_freq.sort_index(axis = 1)

IDF(Inverse Document Frequency)/TF(Term Frequenchy)

import numpy as np

pd.options.display.float_format = '{:.2f}'.format

idf = np.log(len(docs) / doc_freq)

# IDF(Inverse Document Frequency): 각 컬럼의 값이 작을수록 해당하는 칼럼의 단어가 문장에 더 많이 등장했다.
idf

# TF(Term Frequenchy): 특정 문서(행)에서 특정 단어의 등장 횟수

wcs

tf_idf = wcs * idf.iloc[0]

tf_idf

# nomalizaion을 하겠다.

from sklearn.preprocessing import Normalizer

Normalizer(norm = 'l1').fit_transform(tf_idf)

from sklearn.feature_extraction.text import TfidfVectorizer #tf_idf와 관련된 라이브러리 (ex) 문서 유사도 구할 때)

vectorizer = TfidfVectorizer(norm = 'l1', stop_words='english')

sparse = vectorizer.fit_transform(docs)

vectorizer.get_feature_names_out()

sparse.todense()

---

#
from sklearn.datasets import fetch_20newsgroups

twenty_train = fetch_20newsgroups(subset = 'train',
                                  shuffle=True,
                                  random_state=0)

twenty_train.target_names

len(twenty_train.target_names) # 총 20개의 분류 문서가 있음

twenty_train.target_names[15]

from collections import Counter

Counter(twenty_train.target)

print((twenty_train.data[0].split('\n')))

print("\n".join(twenty_train.data[0].split('\n')))

# 문제
pipeline: tf-idf화 시켜서 -> randomforest로 해보기 

from sklearn import datasets
from sklearn.preprocessing import StandardScaler, PolynomialFeatures
from sklearn.svm import SVC

from sklearn.pipeline import make_pipeline
from sklearn.ensemble import RandomForestClassifier

import warnings
warnings.filterwarnings('ignore')

from sklearn import set_config
set_config(display = "diagram")

import pandas as pd

vectorizer = TfidfVectorizer(norm = 'l1', stop_words='english')

pipe = make_pipeline(vectorizer, RandomForestClassifier())

pipe

twenty_train = fetch_20newsgroups(subset = 'train',
                                  shuffle=True,
                                  random_state=0)

twenty_test = fetch_20newsgroups(subset = 'test',
                                  shuffle=True,
                                  random_state=0)

pipe.fit(twenty_train.data, twenty_train.target)

pipe.score(twenty_test.data, twenty_test.target)

# 문제 F1 Score로 다시하기
pipeline: tf-idf화 시켜서 -> randomforest로 해보기 

from sklearn.metrics import f1_score

y_pred = pipe.predict(twenty_test.data)

f1_score(twenty_test.target, y_pred, average='macro')

# GridSearchCV
from sklearn.model_selection import GridSearchCV

# 100의 약수 구하기
n = 100
for i in range(1, 101) :
    if n % i == 0 :
        print(i)

---

# 강사님 풀이

from sklearn.pipeline import make_pipeline
from sklearn.ensemble import RandomForestClassifier

pipe = make_pipeline(TfidfVectorizer(), RandomForestClassifier())

pipe.fit(twenty_train.data, twenty_train.target)

pipe.score(twenty_test.data, twenty_test.target)

from sklearn.metrics import f1_score

pred = pipe.predict(twenty_test.data)

f1_score(twenty_test.target, pred, average='macro')

# GridSearchCV
from sklearn.model_selection import GridSearchCV

pipe

model = RandomForestClassifier()

param_grid = {'randomforestclassifier__criterion' : ['gini', 'entropy', 'log_loss'],
              'randomforestclassifier__max_features' : ['sqrt', 'log2'],
              #'randomforestclassifier__class_weight' : ['balanced', 'balanced_subsample'],
              'tfidfvectorizer__norm' : ['l1','l2']}

grid_model = GridSearchCV(pipe, param_grid, verbose=2, n_jobs= -1)

grid_model.fit(twenty_train.data, twenty_train.target)

grid_model

grid_model.best_estimator_.score(twenty_test.data, twenty_test.target)

pipe.named_steps['randomforestclassifier']

---

# 지예님 코드(문장마다 검사하기)
word_list = []
for sentence in docs :
    word_dict = {}
    for voca in vocabulary2 :
        word_dict[voca] = voca in sentence.split()
    word_list.append(word_dict)

word_list

# comprehension으로 하기


---

docs

sentence

vocabulary2

voca

# 혼자 연습

docs[0].split()

word_dict = {}
for i in docs :
    print(i)
    for j in vocabulary2 :
        if j in i :
            word_dict[i] = {j:"True"}
        else :
            word_dict[i] = {j :"False"}
        

for i in docs :
    print(i)

# 힌트
'cat' in docs[0]

docs[0]

for doc in docs :
    print('cat', 'cat' in doc)

set(['a', 'to', 'the', 'in', 'and', 'are'])

----

# 혼자 연습

for i in range(len(docs)) :
    print(i)
    print(docs[i].split())

for i in range(len(docs)) :
    print(docs[i].split())

set(docs)
result_docs = list(set(docs))

print(result

A = [1, 2, 1, 3, 2, 4, 4, 5, 1]

print(set(A))

result_A = list(set(A))

print(result_A)

docs = ["the cat in the hat",
       "the cow jumped over the moon",
       "the cat mooed and the cow meowed", 
       "the cat said to the cow cow you are not a cat"]
result = [] # 중복 제거된 값들이 들어갈 리스트

for value in docs:
    if value not in result:
        result.append(value)

print(result)

docs = ["the cat in the hat",
       "the cow jumped over the moon",
       "the cat mooed and the cow meowed", 
       "the cat said to the cow cow you are not a cat"]

result1 = dict.fromkeys(docs) # 리스트 값들을 key 로 변경
print(result1)

result2 = list(result1) # list(dict.fromkeys(arr))
print(result2)

import numpy as np

docs = ["the cat in the hat",
       "the cow jumped over the moon",
       "the cat mooed and the cow meowed", 
       "the cat said to the cow cow you are not a cat"]

result_docs = np.unique(docs)

print(result_docs)

---
